refactor(server): split the gateway runtime by concern
Turn gateway_runtime.rs into a directory module. mod.rs keeps the
GatewayRuntime handle, its types, and construction. Each concern
extends the same impl GatewayRuntime from its own file: session
(pairing, sign-in, sign-out, connection), apps (entitlements and
MCP catalog), models (snapshot, sync, route token source), and relay
(shared-app invoke), with the fake-gateway tests in tests.rs.

Move-only: no logic changes, no renamed public items.
