--- crates/tidebreak-server/src/gateway_runtime.rs
@@ -1,4717 +1,0 @@
-//! The server's handle on a signed-in model-gateway session.
-//!
-//! Owns the [`GatewayConnection`] built from the resolved managed policy —
-//! the only gateway source in either direction — hands the router a
-//! per-request token source for the gateway's short-lived `llm` tokens, and
-//! syncs the entitled model list into the stored snapshot so the picker and
-//! model policy work from durable local state while the gateway stays the
-//! live authority at inference time. On an unmanaged profile every surface
-//! here is inert: no connection, no routes, and the sign-in endpoints refuse
-//! with a pointer at the pairing flow.
-
-use std::sync::Arc;
-use std::time::Duration;
-
-use crate::connectors::{
-    is_sign_in_required, CredentialVault, GatewayApp, GatewayAuth, GatewayAuthConfig,
-    GatewayCatalogFetch, GatewayConnection, GatewayOperationSummary, MEMBER_CATALOG_V1,
-    RESOURCE_CONTROL, RESOURCE_LLM,
-};
-use async_trait::async_trait;
-use serde::Serialize;
-use tidebreak_core::{AgentError, Result, SecretProvider, Store};
-use tidebreak_router::{BearerTokenSource, ModelRouteLease};
-use tokio::sync::{Mutex, OwnedRwLockWriteGuard, RwLock};
-
-use crate::providers::{self, CustomModelConfig};
-
-/// How long a browser sign-in may stay pending before it fails.
-const SIGN_IN_TIMEOUT: Duration = Duration::from_secs(300);
-
-/// How often the background loop refreshes the entitled-model snapshot while
-/// a session is connected. A sync is one small GET against the org's own
-/// gateway, so a tight cadence is cheap — this is what bounds how quickly an
-/// admin's entitlement change reaches the picker without a manual refresh.
-const MODEL_SYNC_INTERVAL: Duration = Duration::from_secs(5 * 60);
-
-/// Retry cadence after a failed background sync — short enough that a boot
-/// racing the network coming up doesn't stay stale for a whole interval,
-/// long enough not to hammer an unreachable gateway.
-const MODEL_SYNC_RETRY: Duration = Duration::from_secs(60);
-
-pub(crate) struct GatewayRuntime {
-    store: Arc<dyn Store>,
-    secrets: Arc<dyn SecretProvider>,
-    /// The provisioned-policy home for managed-mode resolution: the sticky
-    /// pairing record, durable beside (not inside) the SQLite profile.
-    provisioned_policy: Arc<dyn crate::managed_policy::ProvisionedPolicySource>,
-    /// The OS authority for managed-mode resolution: a managed profile's
-    /// deployment URL comes from the resolved policy, not the stored row.
-    os_policy: Arc<dyn crate::managed_policy::OsPolicySource>,
-    /// One connection per configured base URL; rebuilt when the URL changes.
-    cached: Mutex<Option<(String, Arc<GatewayConnection>)>>,
-    /// Shared by every request-leg route lease and every local mutation of
-    /// gateway execution authority. Request setup takes a read lease through
-    /// dispatch; catalog sync, sign-out, deprovision, and session replacement
-    /// take the write side. A writer therefore runs either before a leg
-    /// validates and mints its bearer or after that leg has dispatched, never
-    /// in the security-sensitive gap between them. Catalog sync deliberately
-    /// keeps the write side across its fetch so older responses cannot commit
-    /// after newer ones.
-    model_sync: Arc<RwLock<()>>,
-    /// The one in-flight browser sign-in, if any.
-    sign_in: Mutex<SignInProgress>,
-    /// Bumped by every `begin_sign_in` and `sign_out` — and by every pending-
-    /// pairing registration or dismissal; a background exchange task may only
-    /// act while its own generation is still current, so a stale attempt can
-    /// neither clobber a newer one's status, resurrect a signed-out session,
-    /// nor commit a pairing that was dismissed or replaced under it.
-    sign_in_generation: std::sync::atomic::AtomicU64,
-    /// The machine offer last read from the gateway's `/api/v1/meta`, keyed
-    /// by the gateway it was read from so a re-pair reads the new one.
-    /// Only a present offer is cached. An older gateway or a deployment that
-    /// has not published its machine yet is retried while the process runs.
-    machine_offer: Mutex<Option<(String, String)>>,
-    /// A deep-link pairing awaiting the sign-in that is its consent.
-    ///
-    /// Process-ephemeral on purpose: nothing durable exists until the user
-    /// completes a sign-in against this gateway, so an unwanted provision
-    /// link dies with a dismissal or the process. Set only by the desktop
-    /// shell through [`crate::register_pending_pairing`] — the renderer can
-    /// read and dismiss it, never set it or choose its URL.
-    pending_pairing: Mutex<Option<PendingPairing>>,
-    #[cfg(test)]
-    sync_commit_pause: Mutex<Option<Arc<MigrationPause>>>,
-}
-
-#[cfg(test)]
-#[derive(Default)]
-struct MigrationPause {
-    arrived: tokio::sync::Notify,
-    release: tokio::sync::Notify,
-}
-
-/// The shell-registered pairing a sign-in may commit.
-#[derive(Clone)]
-struct PendingPairing {
-    /// Normalized gateway base URL, already held to the connectors contract.
-    base_url: String,
-    /// Carried from the shell's [`crate::PairingHandle`] at registration so
-    /// the commit cannot run without also applying what it decides to the
-    /// MCP servers this process is running.
-    mcp: Arc<crate::mcp_config::McpRuntime>,
-    /// For a re-pairing the user confirmed: the provisioned URL the
-    /// confirmation named, which the commit's compare-and-swap holds the row
-    /// to. `None` for a plain first-time pairing.
-    replaces: Option<String>,
-}
-
-/// Renderer-safe progress of the current sign-in attempt.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, ts_rs::TS)]
-#[serde(tag = "state", rename_all = "snake_case")]
-pub(crate) enum SignInProgress {
-    /// No sign-in is running.
-    Idle,
-    /// The browser flow is open; the renderer should offer this URL.
-    Pending { authorization_url: String },
-    /// The last attempt failed with a bounded, secret-free message.
-    Failed { message: String },
-}
-
-/// Renderer-safe projection of the gateway connection state. Never carries
-/// token material — only what the settings surface displays. `base_url` is
-/// the policy's gateway origin: present exactly when the profile is managed
-/// with a usable URL (the retired `configured`/`enabled` bits collapsed into
-/// its presence).
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, ts_rs::TS)]
-pub(crate) struct GatewayStatus {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[ts(optional)]
-    pub(crate) base_url: Option<String>,
-    pub(crate) signed_in: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[ts(optional)]
-    pub(crate) account_hint: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[ts(optional)]
-    pub(crate) installation_id: Option<String>,
-    pub(crate) model_count: usize,
-    /// The member-catalog contract revision the last model sync read, or
-    /// `None` while unsynced or against a gateway that predates
-    /// `/api/v1/me/catalog`. The settings panel uses its absence (while
-    /// signed in with models) to note that the deployment is older than
-    /// this Tidebreak.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[ts(optional)]
-    pub(crate) member_catalog: Option<String>,
-    pub(crate) sign_in: SignInProgress,
-}
-
-/// The hosted Tidebreak machine this profile's gateway offers, if it offers
-/// one. Absent means the address field stays empty — a gateway that hosts no
-/// machine and a gateway older than the field are the same thing here.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
-pub(crate) struct GatewayMachineOffer {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) url: Option<String>,
-}
-
-/// Renderer-safe list of the connected apps the signed-in user is entitled
-/// to, fetched live from the gateway (never cached: a revoked grant is gone
-/// on the next request).
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, ts_rs::TS)]
-pub(crate) struct GatewayApps {
-    /// False when the connected gateway predates the JSON apps surface; the
-    /// renderer hides the section instead of showing an empty list as "none".
-    pub(crate) supported: bool,
-    pub(crate) apps: Vec<GatewayAppInfo>,
-}
-
-/// One entitled connected app, with the slugs of the MCP endpoints that
-/// aggregate it — the `mcp:<slug>` resources a mount would request.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, ts_rs::TS)]
-pub(crate) struct GatewayAppInfo {
-    pub(crate) id: String,
-    pub(crate) name: String,
-    pub(crate) app_kind: String,
-    pub(crate) enabled: bool,
-    pub(crate) mcp_endpoint_slugs: Vec<String>,
-    /// The gateway's readiness for this app when the member catalog reports
-    /// one: `ready`, `not_connected`, or `authorization_required`. `None`
-    /// against a gateway that predates the catalog — the panel then shows
-    /// no readiness rather than guessing. An unfamiliar value renders as
-    /// not-ready copy, never an error: the set is the gateway's to grow.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[ts(optional)]
-    pub(crate) connection: Option<String>,
-    /// How many live local-app grants bind this gateway app — the same
-    /// "Used by N local apps" line the connected-apps page carries per record,
-    /// so a user can see what a revocation here would break.
-    pub(crate) used_by_app_count: usize,
-}
-
-impl GatewayRuntime {
-    pub(crate) fn new(
-        store: Arc<dyn Store>,
-        secrets: Arc<dyn SecretProvider>,
-        provisioned_policy: Arc<dyn crate::managed_policy::ProvisionedPolicySource>,
-        os_policy: Arc<dyn crate::managed_policy::OsPolicySource>,
-    ) -> Arc<Self> {
-        Arc::new(Self {
-            store,
-            secrets,
-            provisioned_policy,
-            os_policy,
-            cached: Mutex::new(None),
-            model_sync: Arc::new(RwLock::new(())),
-            sign_in: Mutex::new(SignInProgress::Idle),
-            sign_in_generation: std::sync::atomic::AtomicU64::new(0),
-            machine_offer: Mutex::new(None),
-            pending_pairing: Mutex::new(None),
-            #[cfg(test)]
-            sync_commit_pause: Mutex::new(None),
-        })
-    }
-
-    /// The provisioned-policy home this runtime resolves against — the
-    /// pairing write path must commit through the same instance.
-    pub(crate) fn provisioned_policy(
-        &self,
-    ) -> &Arc<dyn crate::managed_policy::ProvisionedPolicySource> {
-        &self.provisioned_policy
-    }
-
-    /// The one policy read every surface here shares.
-    pub(crate) fn policy(&self) -> Result<crate::managed_policy::ManagedPolicy> {
-        crate::managed_policy::resolve(&*self.provisioned_policy, &*self.os_policy)
-    }
-
-    /// The synced model snapshot for the deployment named by managed policy.
-    pub(crate) async fn model_snapshot(&self) -> Result<Option<providers::GatewayModelSnapshot>> {
-        let policy = self.policy()?;
-        providers::gateway_snapshot_for_policy(&*self.store, &policy).await
-    }
-
-    /// Park a shell-validated pairing until a sign-in consents to it,
-    /// replacing any earlier one — the latest link is the one the user acted
-    /// on. Invalidate any in-flight browser flow the same way `sign_out`
-    /// does: an exchange started against a replaced pairing must abandon
-    /// rather than commit it.
-    pub(crate) async fn register_pending_pairing(
-        &self,
-        base_url: String,
-        mcp: Arc<crate::mcp_config::McpRuntime>,
-        replaces: Option<String>,
-    ) {
-        let mut sign_in = self.sign_in.lock().await;
-        self.sign_in_generation
-            .fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-        *sign_in = SignInProgress::Idle;
-        *self.pending_pairing.lock().await = Some(PendingPairing {
-            base_url,
-            mcp,
-            replaces,
-        });
-    }
-
-    /// The pending pairing's gateway URL, for the `/policy` projection.
-    pub(crate) async fn pending_pairing_url(&self) -> Option<String> {
-        self.pending_pairing
-            .lock()
-            .await
-            .as_ref()
-            .map(|pending| pending.base_url.clone())
-    }
-
-    /// Decline the pending pairing: clear it and invalidate any browser flow
-    /// it started. Renderer-reachable, deliberately — declining changes
-    /// nothing durable, so the failure direction is safe. With nothing
-    /// pending it is a strict no-op: the generation must not move, or a
-    /// stray dismiss could abandon a legitimate managed sign-in mid-flight.
-    pub(crate) async fn dismiss_pending_pairing(&self) {
-        let mut sign_in = self.sign_in.lock().await;
-        let mut pending = self.pending_pairing.lock().await;
-        if pending.is_none() {
-            return;
-        }
-        self.sign_in_generation
-            .fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-        *sign_in = SignInProgress::Idle;
-        *pending = None;
-    }
-
-    /// The renderer-facing connection status, derived from policy alone: a
-    /// profile is gateway-connected exactly when managed policy asserts it,
-    /// so an unmanaged profile reads no gateway whatever legacy rows
-    /// persist, and a managed policy whose URL is missing (misconfigured)
-    /// reads none, honestly.
-    pub(crate) async fn status(&self) -> Result<GatewayStatus> {
-        // One policy read for the whole projection: the renderer polls this
-        // every couple of seconds while a sign-in is pending.
-        let policy = crate::managed_policy::resolve(&*self.provisioned_policy, &*self.os_policy)?;
-        let base_url = policy.gateway_url.clone();
-        let credentials = match self.connection_for(&policy).await? {
-            Some(connection) => connection.stored_credentials().await?,
-            None => None,
-        };
-        let snapshot = providers::gateway_snapshot_for_policy(&*self.store, &policy).await?;
-        Ok(GatewayStatus {
-            base_url,
-            signed_in: credentials.is_some(),
-            account_hint: credentials
-                .as_ref()
-                .and_then(|credentials| credentials.account_hint.clone()),
-            installation_id: credentials
-                .as_ref()
-                .map(|credentials| credentials.installation_id.clone()),
-            model_count: snapshot
-                .as_ref()
-                .map(|snapshot| snapshot.models.len())
-                .unwrap_or_default(),
-            member_catalog: snapshot.and_then(|snapshot| snapshot.member_catalog),
-            sign_in: self.sign_in.lock().await.clone(),
-        })
-    }
-
-    /// The hosted machine this profile's gateway offers, for the address
-    /// field on the settings panel.
-    ///
-    /// Read from the gateway's unauthenticated `/api/v1/meta` rather than
-    /// from the provision link: the link fires once at pair time, so a
-    /// profile paired earlier would never see the value and a machine that
-    /// moved would leave a stale one behind. Meta is re-read every boot.
-    ///
-    /// Every failure reads as no offer — an unmanaged profile, a gateway
-    /// older than the field, a gateway that does not answer. The value is a
-    /// hint for a text field and never authorization: attaching still runs
-    /// discovery, which holds the machine to naming this same gateway.
-    pub(crate) async fn offered_machine(&self) -> GatewayMachineOffer {
-        GatewayMachineOffer {
-            url: self.read_offered_machine().await,
-        }
-    }
-
-    /// The offer, memoized per gateway. Only a valid, present offer is
-    /// remembered, so a boot that raced either the network or the gateway
-    /// rollout retries on the next ask. An unmanaged profile that pairs
-    /// mid-session also reads its new gateway rather than a prior absence.
-    async fn read_offered_machine(&self) -> Option<String> {
-        let policy = self.policy().ok()?;
-        if !policy.managed {
-            return None;
-        }
-        let base_url = policy.gateway_url?;
-        let mut memo = self.machine_offer.lock().await;
-        if let Some((read_from, offer)) = memo.as_ref() {
-            if *read_from == base_url {
-                return Some(offer.clone());
-            }
-        }
-        let connection = self.connection_at(base_url.clone()).await.ok()?;
-        let meta = connection.auth().meta().await.ok()?;
-        // The reader sees this in a text box, so hold it to the same URL
-        // rules the connect path enforces: a value that could never be
-        // connected to is worse than an empty field.
-        let offer = meta
-            .tidebreak_machine_url
-            .filter(|url| GatewayAuthConfig::new(url).is_ok());
-        if let Some(offer) = offer.as_ref() {
-            *memo = Some((base_url, offer.clone()));
-        }
-        offer
-    }
-
-    /// The entitled connected apps, fetched live from the gateway with the
-    /// stored session. Managed-only, like the whole sign-in surface; a
-    /// gateway without the JSON apps surface reports `supported: false`.
-    pub(crate) async fn apps(&self, owner: &tidebreak_core::OwnerId) -> Result<GatewayApps> {
-        let connection = self.managed_connection().await?;
-        let Some(apps) = Self::entitled_apps(&connection).await? else {
-            return Ok(GatewayApps {
-                supported: false,
-                apps: Vec::new(),
-            });
-        };
-        // How many local mini-apps currently bind each gateway app: distinct
-        // live grants naming the id. The binding grammar forbids a grant
-        // naming one gateway app twice, so grants and apps count one-to-one.
-        // Best-effort, like every other projection field here — an unreadable
-        // grant table leaves the counts at zero rather than failing the list.
-        let mut used_by: std::collections::BTreeMap<String, usize> =
-            std::collections::BTreeMap::new();
-        match self.store.list_live_app_grants_scoped(owner).await {
-            Ok(grants) => {
-                for grant in grants {
-                    for binding in &grant.bindings {
-                        if let Some(gateway_app) = binding.gateway_app() {
-                            *used_by.entry(gateway_app.to_owned()).or_default() += 1;
-                        }
-                    }
-                }
-            }
-            Err(error) => {
-                tracing::warn!("could not count local apps bound to gateway apps: {error}");
-            }
-        }
-        Ok(GatewayApps {
-            supported: true,
-            apps: apps
-                .into_iter()
-                .map(|mut app| {
-                    app.used_by_app_count = used_by.get(&app.id).copied().unwrap_or_default();
-                    app
-                })
-                .collect(),
-        })
-    }
-
-    /// The entitled connected apps, or `None` when the gateway serves
-    /// neither apps surface.
-    ///
-    /// The member catalog carries per-app readiness the older surface
-    /// cannot express; prefer it, and degrade to `/api/v1/cli/apps`
-    /// (readiness unknown) against a gateway that predates it. Always an
-    /// unconditional fetch: this is the live projection a revoked grant
-    /// must fall out of, so a cached revision would be a lie.
-    async fn entitled_apps(connection: &GatewayConnection) -> Result<Option<Vec<GatewayAppInfo>>> {
-        match connection.catalog(None).await? {
-            GatewayCatalogFetch::Fresh { catalog, .. } => Ok(Some(
-                catalog
-                    .apps
-                    .into_iter()
-                    .map(|app| GatewayAppInfo {
-                        used_by_app_count: 0,
-                        id: app.id,
-                        name: app.name,
-                        app_kind: app.app_kind,
-                        enabled: app.enabled,
-                        mcp_endpoint_slugs: app.mcp_endpoint_slugs,
-                        connection: Some(app.connection),
-                    })
-                    .collect(),
-            )),
-            GatewayCatalogFetch::NotModified => Err(AgentError::msg(
-                "gateway answered an unconditional catalog fetch with 304",
-            )),
-            GatewayCatalogFetch::Unsupported => Ok(connection.apps().await?.map(|apps| {
-                apps.into_iter()
-                    .map(|app| GatewayAppInfo {
-                        used_by_app_count: 0,
-                        id: app.id,
-                        name: app.name,
-                        app_kind: app.app_kind,
-                        enabled: app.enabled,
-                        mcp_endpoint_slugs: app.mcp_endpoint_slugs,
-                        connection: None,
-                    })
-                    .collect()
-            })),
-        }
-    }
-
-    /// Mount newly entitled gateway MCP endpoints into the configured server
-    /// set — mount-by-default for a managed profile, where the organization
-    /// already curated the entitlements.
-    ///
-    /// The entitlement source is [`entitled_apps`](Self::entitled_apps) — the
-    /// same read the settings panel lists, so server and UI cannot disagree
-    /// about what is entitled. Endpoints the user explicitly unmounted are
-    /// remembered by the MCP runtime and never re-mounted here; a repeat
-    /// reconcile with no new entitlements changes nothing. Every state where
-    /// a reconcile cannot run — unmanaged profile, misconfigured policy, no
-    /// session for the policy's deployment, a gateway predating the apps
-    /// surface — is "nothing to do", not an error, so callers may run this on
-    /// every trigger without gating.
-    pub(crate) async fn reconcile_endpoint_mounts(
-        &self,
-        mcp: &crate::mcp_config::McpRuntime,
-    ) -> Result<()> {
-        let policy = self.policy()?;
-        let Some(connection) = self.connection_for(&policy).await? else {
-            return Ok(());
-        };
-        if connection.stored_credentials().await?.is_none() {
-            return Ok(());
-        }
-        let Some(apps) = Self::entitled_apps(&connection).await? else {
-            return Ok(());
-        };
-        let entitled: Vec<String> = apps
-            .into_iter()
-            .flat_map(|app| app.mcp_endpoint_slugs)
-            .collect::<std::collections::BTreeSet<_>>()
-            .into_iter()
-            .collect();
-        if entitled.is_empty() {
-            return Ok(());
-        }
-        if mcp.auto_mount_gateway_endpoints(&entitled).await? {
-            tracing::info!("auto-mounted newly entitled gateway MCP endpoints");
-        }
-        Ok(())
-    }
-
-    /// The gateway's deployment URL and the apps the signed-in user is
-    /// entitled to, or `None` when this profile cannot answer that question.
-    ///
-    /// The non-refusing twin of [`apps`](Self::apps): every state where there
-    /// is nothing to read — unmanaged profile, misconfigured policy, no
-    /// session for the policy's deployment, a session the gateway has since
-    /// revoked, a gateway predating the apps surface — answers `None` rather
-    /// than an error, exactly as [`reconcile_endpoint_mounts`] treats them as
-    /// "nothing to do". Callers use this where a gateway that cannot answer
-    /// must degrade (an authoring roster) or fail closed (a fingerprint read),
-    /// never fault.
-    ///
-    /// The URL is the resolved policy's, the same string every other gateway
-    /// identity is stamped with — so a fingerprint taken from it moves when
-    /// and only when the profile is re-paired.
-    pub(crate) async fn entitled_apps_if_signed_in(
-        &self,
-    ) -> Result<Option<(String, Vec<GatewayApp>)>> {
-        let policy = self.policy()?;
-        let Some(base_url) = policy.gateway_url.clone().filter(|_| policy.managed) else {
-            return Ok(None);
-        };
-        let Some(connection) = self.connection_for(&policy).await? else {
-            return Ok(None);
-        };
-        if connection.stored_credentials().await?.is_none() {
-            return Ok(None);
-        }
-        match connection.apps().await {
-            Ok(Some(apps)) => Ok(Some((base_url, apps))),
-            Ok(None) => Ok(None),
-            // A revoked or expired session is "signed out", not a fault: the
-            // stored blob outlives the gateway's own session record.
-            Err(error) if is_sign_in_required(&error) => Ok(None),
-            Err(error) => Err(error),
-        }
-    }
-
-    /// One entitled app's declared operation catalog, under the same gating as
-    /// [`entitled_apps_if_signed_in`](Self::entitled_apps_if_signed_in): a
-    /// profile that cannot read it answers `None`, never an error.
-    pub(crate) async fn app_catalog(
-        &self,
-        app_id: &str,
-    ) -> Result<Option<Vec<GatewayOperationSummary>>> {
-        let policy = self.policy()?;
-        let Some(connection) = self.connection_for(&policy).await? else {
-            return Ok(None);
-        };
-        if connection.stored_credentials().await?.is_none() {
-            return Ok(None);
-        }
-        match connection.app_operations(app_id).await {
-            Ok(catalog) => Ok(catalog),
-            Err(error) if is_sign_in_required(&error) => Ok(None),
-            Err(error) => Err(error),
-        }
-    }
-
-    /// Every bindable gateway connected app with the operation ids its
-    /// catalog declares — the one live read both the `create_app` roster and
-    /// the authoring door work from.
-    ///
-    /// `None` is "this profile has no gateway session to ask", which the door
-    /// turns into a teachable refusal; `Some(vec![])` is a session that
-    /// answered with nothing bindable. An app whose catalog cannot be read is
-    /// left out: nothing about it could be pinned, so listing it would only
-    /// invite a refusal one step later.
-    pub(crate) async fn app_roster(&self) -> Option<Vec<crate::mcp_config::GatewayRosterApp>> {
-        let entitled = match self.entitled_apps_if_signed_in().await {
-            Ok(Some((_, entitled))) => entitled,
-            Ok(None) => return None,
-            Err(error) => {
-                tracing::warn!("could not read the gateway's entitled apps: {error}");
-                return None;
-            }
-        };
-        let mut roster = Vec::new();
-        for app in entitled.into_iter().filter(|app| app.enabled) {
-            match self.app_catalog(&app.id).await {
-                Ok(Some(operations)) => roster.push(crate::mcp_config::GatewayRosterApp {
-                    id: app.id,
-                    name: app.name,
-                    operation_ids: operations
-                        .into_iter()
-                        .map(|operation| operation.operation_id)
-                        .collect(),
-                }),
-                Ok(None) => {}
-                Err(error) => {
-                    tracing::warn!(
-                        gateway_app = %app.id,
-                        "could not read a gateway app's operation catalog: {error}"
-                    );
-                }
-            }
-        }
-        Some(roster)
-    }
-
-    /// Start a browser sign-in and return the URL to open.
-    ///
-    /// A pending pairing wins the target: the sign-in runs against the
-    /// pairing's gateway, and a successful exchange is what commits the
-    /// provision — the sign-in the user chose to complete is the pairing's
-    /// consent, so nothing durable exists until it succeeds. That holds on a
-    /// managed profile too: a pending pairing can only exist there through
-    /// the shell's confirmed re-pair flow, never a bare deep link, so
-    /// honoring it is honoring that confirmation. With nothing pending, a
-    /// managed profile's sign-in targets the policy's gateway, and an
-    /// unmanaged one keeps the legible refusal.
-    ///
-    /// The exchange completes in a background task: on success the session is
-    /// stored (after any pairing commit), the entitled models synced, and the
-    /// entitled MCP endpoints auto-mounted into `mcp`; on failure the status
-    /// surface carries the bounded error until the next attempt.
-    pub(crate) async fn begin_sign_in(
-        self: &Arc<Self>,
-        mcp: Arc<crate::mcp_config::McpRuntime>,
-    ) -> Result<String> {
-        let policy = self.policy()?;
-        let pairing = self.pending_pairing.lock().await.clone();
-        let connection = match &pairing {
-            Some(pending) => self.connection_at(pending.base_url.clone()).await?,
-            None => self.connection_at(require_managed(&policy)?).await?,
-        };
-        let pending = connection.auth().start_sign_in().await?;
-        let authorization_url = pending.authorization_url().to_string();
-        let generation = {
-            let mut sign_in = self.sign_in.lock().await;
-            let generation = self
-                .sign_in_generation
-                .fetch_add(1, std::sync::atomic::Ordering::SeqCst)
-                + 1;
-            *sign_in = SignInProgress::Pending {
-                authorization_url: authorization_url.clone(),
-            };
-            generation
-        };
-
-        let runtime = self.clone();
-        tokio::spawn(async move {
-            let finished = pending.finish(SIGN_IN_TIMEOUT).await;
-            let session = match finished {
-                Ok(session) => session,
-                Err(error) => {
-                    let mut sign_in = runtime.sign_in.lock().await;
-                    if runtime
-                        .sign_in_generation
-                        .load(std::sync::atomic::Ordering::SeqCst)
-                        == generation
-                    {
-                        *sign_in = SignInProgress::Failed {
-                            message: error.to_string(),
-                        };
-                    }
-                    return;
-                }
-            };
-
-            // Lock order for every locally controlled authority mutation is
-            // model authority -> pairing (when policy may move) -> sign-in
-            // state. The write lease waits for already-authorized request legs
-            // to dispatch, then excludes new ones through the policy commit,
-            // old-session retirement, and new-session store. Acquiring the
-            // pairing lock before sign-in state also matches registration and
-            // deprovision, avoiding the former sign-in/pairing inversion.
-            let _authority = runtime.lock_model_authority_mutation().await;
-            let _pairing = crate::pairing::lock_pairing_mutation().await;
-            let mut sign_in = runtime.sign_in.lock().await;
-            if runtime
-                .sign_in_generation
-                .load(std::sync::atomic::Ordering::SeqCst)
-                != generation
-            {
-                return;
-            }
-
-            // A pairing commits before its session persists: if the provision
-            // cannot be written (an MDM push claimed the profile mid-flow), no
-            // session lands on a profile the pairing's gateway does not manage.
-            let committed = match &pairing {
-                Some(pending) => runtime.commit_pairing_locked(pending).await,
-                None => Ok(()),
-            };
-            let stored = match committed {
-                Ok(()) => connection.store_session(&session).await,
-                Err(error) => Err(error),
-            };
-            let stored = match stored {
-                Ok(()) => {
-                    *sign_in = SignInProgress::Idle;
-                    true
-                }
-                Err(error) => {
-                    *sign_in = SignInProgress::Failed {
-                        message: error.to_string(),
-                    };
-                    false
-                }
-            };
-            drop(sign_in);
-            drop(_pairing);
-            drop(_authority);
-
-            if !stored {
-                return;
-            }
-            // These are post-commit refreshes, not authority mutations. Run
-            // them after releasing the writer so healthy inference is not
-            // serialized behind model and endpoint network reads.
-            if let Err(error) = runtime.sync_models().await {
-                tracing::warn!(
-                    "gateway model sync after sign-in failed \
-                     (the background sync will retry): {}",
-                    error.message()
-                );
-            }
-            if let Err(error) = runtime.reconcile_endpoint_mounts(&mcp).await {
-                tracing::warn!(
-                    "gateway endpoint auto-mount after sign-in failed \
-                     (the background sync will retry): {error}"
-                );
-            }
-            mcp.refresh_connected_app_roster().await;
-        });
-        Ok(authorization_url)
-    }
-
-    /// Commit the pairing a finishing sign-in consented to, then clear it.
-    /// Runs from the exchange task with the sign-in state lock held, so it
-    /// cannot interleave with a dismissal or re-registration.
-    async fn commit_pairing_locked(&self, pending: &PendingPairing) -> Result<()> {
-        crate::pairing::commit_signed_in_pairing_locked(
-            &*self.provisioned_policy,
-            &*self.os_policy,
-            self.secrets.clone(),
-            &pending.mcp,
-            &pending.base_url,
-            pending.replaces.as_deref(),
-        )
-        .await?;
-        *self.pending_pairing.lock().await = None;
-        Ok(())
-    }
-
-    /// Take the exclusive side of the request-leg authority fence.
-    ///
-    /// Callers that also need the pairing or sign-in locks must take them
-    /// after this guard, in that order. The owned guard lets pairing code hold
-    /// the fence across its compare-and-swap and session retirement without
-    /// exposing the lock itself.
-    pub(crate) async fn lock_model_authority_mutation(&self) -> OwnedRwLockWriteGuard<()> {
-        self.model_sync.clone().write_owned().await
-    }
-
-    #[cfg(test)]
-    pub(crate) async fn commit_signed_in_pairing_for_test(
-        &self,
-        mcp: &crate::mcp_config::McpRuntime,
-        base_url: &str,
-        replaces: Option<&str>,
-    ) -> Result<()> {
-        let _authority = self.lock_model_authority_mutation().await;
-        let _pairing = crate::pairing::lock_pairing_mutation().await;
-        let _sign_in = self.sign_in.lock().await;
-        crate::pairing::commit_signed_in_pairing_locked(
-            &*self.provisioned_policy,
-            &*self.os_policy,
-            self.secrets.clone(),
-            mcp,
-            base_url,
-            replaces,
-        )
-        .await
-    }
-
-    /// Revoke the session (best-effort at the gateway), clear local state, and
-    /// drop the synced model snapshot. Managed-only, like sign-in.
-    pub(crate) async fn sign_out(&self) -> Result<()> {
-        // Authority is always outermost: an already-authorized request leg
-        // dispatches before this writer proceeds, while a later leg observes
-        // the cleared session and snapshot. The sign-in state lock nests
-        // inside it, matching sign-in completion.
-        let _model_sync = self.lock_model_authority_mutation().await;
-        // Take the state lock for the whole operation and invalidate any
-        // pending browser flow before revoking anything: an exchange that
-        // completes during the revoke round-trip serializes behind this lock
-        // and then observes the bump, so it abandons instead of re-saving
-        // the session it just minted.
-        let mut sign_in = self.sign_in.lock().await;
-        self.sign_in_generation
-            .fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-        let policy = crate::managed_policy::resolve(&*self.provisioned_policy, &*self.os_policy)?;
-        let base_url = require_managed(&policy)?;
-        self.connection_at(base_url.clone())
-            .await?
-            .sign_out()
-            .await?;
-        {
-            // Serialize the snapshot clear with the other snapshot writers so
-            // it cannot land inside one of their recheck-and-write windows.
-            // Lock order matches the sign-in task's sync path: the sign-in
-            // state lock is already held, the snapshot lock nests inside it.
-            let _lock = providers::GATEWAY_STATE_WRITES.lock().await;
-            if !providers::gateway_models(&*self.store, &policy, None)
-                .await?
-                .is_empty()
-            {
-                providers::write_gateway_snapshot(
-                    &*self.store,
-                    &providers::GatewayModelSnapshot {
-                        gateway_url: base_url,
-                        installation_id: None,
-                        models: Vec::new(),
-                        model_protocols: Default::default(),
-                        model_reasoning_efforts: Default::default(),
-                        member_catalog: None,
-                        catalog_etag: None,
-                    },
-                )
-                .await?;
-            }
-        }
-        *sign_in = SignInProgress::Idle;
-        Ok(())
-    }
-
-    /// Invalidate any in-flight browser sign-in and drop the pending pairing
-    /// — the disconnect epilogue. Same discipline as sign-out: an exchange
-    /// that completes afterwards serializes behind the state lock, observes
-    /// the generation bump, and abandons rather than re-saving the session
-    /// it just minted.
-    pub(crate) async fn abandon_sign_in_and_pairing(&self) {
-        let mut sign_in = self.sign_in.lock().await;
-        self.sign_in_generation
-            .fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-        *self.pending_pairing.lock().await = None;
-        *sign_in = SignInProgress::Idle;
-    }
-
-    /// Retire the stored session against the current resolved policy: after
-    /// a deprovision the profile is unmanaged, so [`sign_out`](Self::sign_out)
-    /// (managed-only) cannot run, but the keychain session must still be
-    /// revoked (best-effort) and cleared (unconditionally). Thin wrapper so
-    /// callers without the secrets handle can reach
-    /// [`retire_superseded_gateway_session`].
-    pub(crate) async fn retire_session_for_current_policy(
-        &self,
-        _authority: &OwnedRwLockWriteGuard<()>,
-    ) -> Result<()> {
-        let policy = self.policy()?;
-        retire_superseded_gateway_session(self.secrets.clone(), &policy).await
-    }
-
-    /// The connection for the policy's gateway, or `None` when the profile is
-    /// unmanaged (or the managed policy is misconfigured).
-    ///
-    /// The deployment comes from the resolved policy and nowhere else. The
-    /// retired provider row was renderer-writable while unmanaged, so
-    /// honoring it here would let a pre-provisioning write redirect sign-in
-    /// and every minted bearer; it is never read.
-    pub(crate) async fn connection(&self) -> Result<Option<Arc<GatewayConnection>>> {
-        let policy = crate::managed_policy::resolve(&*self.provisioned_policy, &*self.os_policy)?;
-        self.connection_for(&policy).await
-    }
-
-    /// [`connection`](Self::connection) against an already-resolved policy, for
-    /// callers that have one in hand.
-    async fn connection_for(
-        &self,
-        policy: &crate::managed_policy::ManagedPolicy,
-    ) -> Result<Option<Arc<GatewayConnection>>> {
-        let Some(base_url) = policy.gateway_url.clone().filter(|_| policy.managed) else {
-            return Ok(None);
-        };
-        Ok(Some(self.connection_at(base_url).await?))
-    }
-
-    /// The connection for the managed gateway, refusing legibly when the
-    /// profile is unmanaged: the sign-in surface (sign-in, sign-out, apps,
-    /// model sync) exists only under managed policy.
-    async fn managed_connection(&self) -> Result<Arc<GatewayConnection>> {
-        let policy = crate::managed_policy::resolve(&*self.provisioned_policy, &*self.os_policy)?;
-        let base_url = require_managed(&policy)?;
-        self.connection_at(base_url).await
-    }
-
-    /// The cached connection for `base_url`, rebuilt when the URL changes.
-    async fn connection_at(&self, base_url: String) -> Result<Arc<GatewayConnection>> {
-        let mut cached = self.cached.lock().await;
-        if let Some((url, connection)) = cached.as_ref() {
-            if *url == base_url {
-                return Ok(connection.clone());
-            }
-        }
-        let auth_config = GatewayAuthConfig::new(&base_url)?;
-        let connection = Arc::new(GatewayConnection::new(
-            GatewayAuth::new(auth_config)?,
-            CredentialVault::new(self.secrets.clone()),
-        ));
-        *cached = Some((base_url, connection.clone()));
-        Ok(connection)
-    }
-
-    /// A router token source, when policy names a gateway and a session for
-    /// that deployment is stored. `None` keeps the gateway route out of the
-    /// router entirely — including on unmanaged profiles and when the stored
-    /// session belongs to a different gateway than the policy URL.
-    pub(crate) async fn route_token_source(self: &Arc<Self>) -> Option<Arc<dyn BearerTokenSource>> {
-        let connection = self.connection().await.ok().flatten()?;
-        let credentials = connection.stored_credentials().await.ok().flatten()?;
-        Some(Arc::new(GatewayTokenSource {
-            connection,
-            installation_id: credentials.installation_id,
-            store: self.store.clone(),
-            provisioned_policy: self.provisioned_policy.clone(),
-            os_policy: self.os_policy.clone(),
-            model_sync: self.model_sync.clone(),
-        }))
-    }
-
-    /// Fetch the entitled models and persist them as the stored snapshot,
-    /// stamped with the deployment they came from. Managed-only.
-    ///
-    /// Returns how many models are entitled. The persisted snapshot drives the
-    /// picker and model policy; entitlement itself stays live at the gateway,
-    /// which refuses a revoked model at inference time regardless of what is
-    /// cached here.
-    pub(crate) async fn sync_models(
-        &self,
-    ) -> std::result::Result<usize, crate::error::ServerError> {
-        // The lock covers the fetch, not merely the write tail. Otherwise two
-        // callers can start from the same ETag and commit in response order,
-        // allowing an older response to replace a newer catalog.
-        let _sync = self.model_sync.write().await;
-        let policy = crate::managed_policy::resolve(&*self.provisioned_policy, &*self.os_policy)?;
-        let base_url = require_managed(&policy)?;
-        let connection = self.connection_at(base_url.clone()).await?;
-        // Populate the control-token cache before taking the credential
-        // fingerprint. The catalog call then reuses that token instead of
-        // rotating this session's refresh token after we captured it.
-        connection.access_token(RESOURCE_CONTROL).await?;
-        let session = connection.stored_credentials().await?.ok_or_else(|| {
-            crate::error::ServerError::conflict_kind(
-                "gateway_changed",
-                "the model gateway session changed during model sync",
-            )
-        })?;
-        // The stored snapshot's ETag makes an unchanged catalog a 304: the
-        // background loop runs this every few minutes, and most ticks change
-        // nothing. Read before the row lock — like the fetch itself, the
-        // conditional round-trip must not stall other writers.
-        let held = providers::gateway_snapshot_for_policy(&*self.store, &policy).await?;
-        let held_etag = held
-            .as_ref()
-            .filter(|snapshot| {
-                snapshot.installation_id.as_deref() == Some(&session.installation_id)
-            })
-            .and_then(|snapshot| snapshot.catalog_etag.clone());
-        // Fetch the full entitled set before the row lock is taken. The
-        // member catalog is preferred: one envelope, merged protocols per
-        // model, and the alias list that matches curated rows. A gateway
-        // that predates it degrades to the per-protocol
-        // `/api/v1/cli/models` read.
-        let mut model_protocols = std::collections::BTreeMap::new();
-        let mut model_reasoning_efforts = std::collections::BTreeMap::new();
-        let mut member_catalog = None;
-        let mut catalog_etag = None;
-        let models: Vec<CustomModelConfig> = match connection.catalog(held_etag.as_deref()).await? {
-            GatewayCatalogFetch::NotModified => {
-                // Nothing moved since the held revision; the snapshot the
-                // policy already honors stays as it is. Recheck the policy and
-                // session before accepting the response so sign-out, session
-                // replacement, and MDM repoints fence every sync outcome.
-                let count = held
-                    .as_ref()
-                    .map(|snapshot| snapshot.models.len())
-                    .unwrap_or_default();
-                self.pause_sync_commit_for_test().await;
-                let _lock = providers::GATEWAY_STATE_WRITES.lock().await;
-                self.recheck_sync_context(&base_url, &session).await?;
-                return Ok(count);
-            }
-            GatewayCatalogFetch::Fresh { catalog, etag } => {
-                member_catalog = Some(MEMBER_CATALOG_V1.to_owned());
-                catalog_etag = etag;
-                let converted = providers::member_catalog_models(catalog);
-                model_protocols = converted.model_protocols;
-                model_reasoning_efforts = converted.model_reasoning_efforts;
-                converted.models
-            }
-            GatewayCatalogFetch::Unsupported => connection
-                .models(None)
-                .await?
-                .into_iter()
-                .filter_map(|model| {
-                    let Some(protocol) = providers::GatewayModelProtocol::parse(&model.protocol)
-                    else {
-                        tracing::debug!(
-                            "gateway model sync skipped a model with an unsupported inference protocol"
-                        );
-                        return None;
-                    };
-                    let id = model.id;
-                    model_protocols.insert(id.clone(), protocol);
-                    Some(CustomModelConfig {
-                        id,
-                        display_name: Some(model.name),
-                        // Carried so a deployment-aliased id can still be matched to
-                        // its curated row; gateways older than the field send none.
-                        upstream_id: model.upstream_id,
-                        aliases: Vec::new(),
-                        context_window: providers::clamp_u32(model.context_window, 32_768),
-                        max_output_tokens: providers::clamp_u32(model.max_output_tokens, 4_096),
-                        input_modalities: vec![crate::model_registry::InputModality::Text],
-                        supports_reasoning: false,
-                        reasoning_efforts: Vec::new(),
-                    })
-                })
-                .collect(),
-        };
-        // The gateway is trusted for entitlements, not for shapes: the synced
-        // set is held to the same bounds as user-entered custom models.
-        providers::validate_custom_models(&models).map_err(|error| {
-            AgentError::config(format!("gateway model sync rejected: {error:?}"))
-        })?;
-        let count = models.len();
-        self.pause_sync_commit_for_test().await;
-        let _lock = providers::GATEWAY_STATE_WRITES.lock().await;
-        // The fetch ran outside the lock, so the policy authority (an MDM
-        // push) may have re-pointed the deployment while it was in flight.
-        // Re-resolve under the lock and refuse to stamp a snapshot the new
-        // policy never entitled.
-        self.recheck_sync_context(&base_url, &session).await?;
-        providers::write_gateway_snapshot(
-            &*self.store,
-            &providers::GatewayModelSnapshot {
-                gateway_url: base_url.clone(),
-                installation_id: Some(session.installation_id.clone()),
-                models,
-                model_protocols,
-                model_reasoning_efforts,
-                member_catalog,
-                catalog_etag,
-            },
-        )
-        .await?;
-        Ok(count)
-    }
-
-    fn recheck_sync_policy(
-        &self,
-        base_url: &str,
-    ) -> std::result::Result<crate::managed_policy::ManagedPolicy, crate::error::ServerError> {
-        let policy = crate::managed_policy::resolve(&*self.provisioned_policy, &*self.os_policy)?;
-        if policy.gateway_url.as_deref() != Some(base_url) {
-            // A benign, retryable race — the deployment was re-pointed while
-            // the fetch was in flight — not an internal fault. The stable
-            // kind lets clients branch on it, as with `managed_profile`.
-            return Err(crate::error::ServerError::conflict_kind(
-                "gateway_changed",
-                "the model gateway configuration changed during model sync",
-            ));
-        }
-        Ok(policy)
-    }
-
-    /// Revalidate both authorities that made a catalog response admissible.
-    /// The policy URL alone does not distinguish sign-out or replacement by a
-    /// different session at the same deployment.
-    async fn recheck_sync_context(
-        &self,
-        base_url: &str,
-        expected_session: &crate::connectors::GatewayCredentials,
-    ) -> std::result::Result<crate::managed_policy::ManagedPolicy, crate::error::ServerError> {
-        let policy = self.recheck_sync_policy(base_url)?;
-        let current = CredentialVault::new(self.secrets.clone()).load().await?;
-        let unchanged = current.as_ref().is_some_and(|credentials| {
-            credentials.matches_base_url(base_url)
-                && same_gateway_session(credentials, expected_session)
-        });
-        if !unchanged {
-            return Err(crate::error::ServerError::conflict_kind(
-                "gateway_changed",
-                "the model gateway session changed during model sync",
-            ));
-        }
-        Ok(policy)
-    }
-
-    #[cfg(test)]
-    async fn pause_sync_commit_for_test(&self) {
-        let pause = self.sync_commit_pause.lock().await.take();
-        if let Some(pause) = pause {
-            pause.arrived.notify_one();
-            pause.release.notified().await;
-        }
-    }
-
-    #[cfg(not(test))]
-    async fn pause_sync_commit_for_test(&self) {}
-
-    /// Keep the entitled-model snapshot fresh without a manual refresh: sync
-    /// once immediately (the boot case) and then on a long interval, for as
-    /// long as the process runs. Every state where a sync cannot run —
-    /// unmanaged profile, misconfigured policy, no session for the policy's
-    /// deployment — is "nothing to do", not an error: the loop waits for the
-    /// state to change rather than exiting, because sign-in, pairing, and MDM
-    /// pushes can all happen at any time.
-    ///
-    /// The same tick reconciles the entitled MCP endpoint mounts into `mcp`:
-    /// the boot-with-stored-session case lands on the immediate first tick,
-    /// and an admin's new entitlement reaches the tool surface within the
-    /// sync interval. Each half fails independently — a failed entitlement
-    /// fetch degrades to "no reconcile this tick", never touching the
-    /// configuration.
-    pub(crate) async fn sync_models_periodically(
-        self: Arc<Self>,
-        mcp: Arc<crate::mcp_config::McpRuntime>,
-    ) {
-        // One warning per outage, not one per retry: the failure state can
-        // legitimately persist for hours on an offline laptop.
-        let mut warned = false;
-        let mut mount_warned = false;
-        loop {
-            let delay = match self.sync_models_if_connected().await {
-                Ok(synced) => {
-                    if let Some(count) = synced {
-                        tracing::debug!("background gateway model sync: {count} models entitled");
-                    }
-                    warned = false;
-                    MODEL_SYNC_INTERVAL
-                }
-                Err(error) => {
-                    let message = error.message();
-                    if warned {
-                        tracing::debug!("background gateway model sync still failing: {message}");
-                    } else {
-                        tracing::warn!(
-                            "background gateway model sync failed (will retry): {message}"
-                        );
-                        warned = true;
-                    }
-                    MODEL_SYNC_RETRY
-                }
-            };
-            match self.reconcile_endpoint_mounts(&mcp).await {
-                Ok(()) => mount_warned = false,
-                Err(error) if mount_warned => {
-                    tracing::debug!("gateway endpoint auto-mount still failing: {error}");
-                }
-                Err(error) => {
-                    tracing::warn!("gateway endpoint auto-mount failed (will retry): {error}");
-                    mount_warned = true;
-                }
-            }
-            tokio::time::sleep(delay).await;
-        }
-    }
-
-    /// One background sync attempt: `Ok(None)` when there is nothing to sync,
-    /// `Ok(Some(count))` after a successful sync. The connected check mirrors
-    /// [`status`](Self::status): a stored session counts only when it belongs
-    /// to the policy's deployment.
-    async fn sync_models_if_connected(
-        &self,
-    ) -> std::result::Result<Option<usize>, crate::error::ServerError> {
-        let policy = self.policy()?;
-        let Some(connection) = self.connection_for(&policy).await? else {
-            return Ok(None);
-        };
-        if connection.stored_credentials().await?.is_none() {
-            return Ok(None);
-        }
-        Ok(Some(self.sync_models().await?))
-    }
-}
-
-/// Attestation-context key for connect-time MCP traffic — `initialize` and
-/// `tools/list` belong to no chat, but a gateway-attested endpoint refuses
-/// any token minted without a context, so the handshake rides a shared one.
-/// Not a UUID, so it can never collide with a chat key.
-const MCP_CONNECT_CONTEXT_KEY: &str = "mcp:connect";
-
-#[async_trait]
-impl crate::mcp_config::GatewayEndpoints for GatewayRuntime {
-    /// Resolve a gateway MCP endpoint from the signed-in session: its URL
-    /// under the configured base, and a fresh `mcp:<slug>` bearer minted (or
-    /// served from cache) inside the connector's rotation lock. The bearer
-    /// carries the shared connect context, which is what lets an attested
-    /// endpoint accept the handshake and list its tools.
-    async fn endpoint(&self, slug: &str) -> Result<crate::mcp_config::GatewayEndpointAccess> {
-        let connection = self
-            .connection()
-            .await?
-            .ok_or_else(|| AgentError::config("no model gateway is configured"))?;
-        Ok(crate::mcp_config::GatewayEndpointAccess {
-            url: connection.mcp_endpoint_url(slug)?,
-            bearer_token: connection
-                .attested_mcp_access_token(slug, MCP_CONNECT_CONTEXT_KEY)
-                .await?,
-        })
-    }
-
-    /// A `tools/call` from a chat rides a token minted inside that chat's
-    /// attestation context — the same context the chat's inference tokens
-    /// carry — so an attested endpoint can match the call against the
-    /// observation that inference recorded. Direct endpoints ignore the
-    /// context, so every gateway mount dispatches this way.
-    async fn call_bearer(&self, slug: &str, chat: tidebreak_core::id::ChatId) -> Result<String> {
-        let connection = self
-            .connection()
-            .await?
-            .ok_or_else(|| AgentError::config("no model gateway is configured"))?;
-        connection
-            .attested_mcp_access_token(slug, &chat.to_string())
-            .await
-    }
-
-    /// The gateway apps the `create_app` roster lists, read live through the
-    /// stored session. Best-effort throughout: no session, an unreachable
-    /// gateway, or a gateway predating either read all answer empty, which
-    /// renders as an absent roster section rather than a failed registry
-    /// rebuild.
-    async fn entitled_app_catalogs(&self) -> Vec<crate::mcp_config::GatewayRosterApp> {
-        self.app_roster().await.unwrap_or_default()
-    }
-}
-
-#[async_trait]
-impl crate::connected_apps::GatewayCatalogSource for GatewayRuntime {
-    /// Resolve the requested gateway apps against the live session: one
-    /// entitled-apps read, then one catalog read per requested id that the
-    /// session actually reaches.
-    ///
-    /// Fail-closed and quiet. A profile that cannot answer at all — unmanaged,
-    /// signed out, a gateway predating either read — answers `None`, and any
-    /// single app that is disabled, unentitled, or catalog-less is simply
-    /// absent from the map. Both readings make a grant naming it stale rather
-    /// than matching a fingerprint over a catalog nobody read.
-    async fn gateway_app_catalogs(
-        &self,
-        needed: &std::collections::BTreeSet<String>,
-    ) -> Option<(
-        String,
-        std::collections::BTreeMap<String, crate::connected_apps::GatewayAppCatalog>,
-    )> {
-        let (base_url, entitled) = match self.entitled_apps_if_signed_in().await {
-            Ok(Some(answered)) => answered,
-            Ok(None) => return None,
-            Err(error) => {
-                tracing::warn!("could not read the gateway's entitled apps: {error}");
-                return None;
-            }
-        };
-        let mut catalogs = std::collections::BTreeMap::new();
-        for app in entitled
-            .into_iter()
-            .filter(|app| app.enabled && needed.contains(&app.id))
-        {
-            match self.app_catalog(&app.id).await {
-                Ok(Some(operations)) => {
-                    catalogs.insert(
-                        app.id,
-                        crate::connected_apps::GatewayAppCatalog {
-                            name: app.name,
-                            operation_ids: operations
-                                .into_iter()
-                                .map(|operation| operation.operation_id)
-                                .collect(),
-                        },
-                    );
-                }
-                Ok(None) => {}
-                Err(error) => {
-                    tracing::warn!(
-                        gateway_app = %app.id,
-                        "could not read a gateway app's operation catalog: {error}"
-                    );
-                }
-            }
-        }
-        Some((base_url, catalogs))
-    }
-}
-
-/// The production gateway relay: the one gateway runtime plus the
-/// registration seam, wired behind [`GatewayInvokeDispatcher`].
-///
-/// Two gates before any operation is relayed, both fail-closed: the profile
-/// must hold a session at the deployment policy names, and the app must be
-/// registered there — which the registration seam establishes on the spot
-/// when it is not, so a granted app self-heals into a servable one on its
-/// first call.
-pub(crate) struct GatewayRelayDispatcher {
-    runtime: Arc<GatewayRuntime>,
-    drafts: Arc<dyn crate::connected_apps::GatewayDraftSource>,
-}
-
-/// The relay as production assembles it: over the process's one gateway
-/// runtime and the store-backed registration registry.
-pub(crate) fn gateway_relay_dispatcher(
-    runtime: Arc<GatewayRuntime>,
-    drafts: Arc<dyn crate::connected_apps::GatewayDraftSource>,
-) -> Arc<dyn crate::connected_apps::GatewayInvokeDispatcher> {
-    Arc::new(GatewayRelayDispatcher { runtime, drafts })
-}
-
-impl GatewayRelayDispatcher {
-    /// One relay attempt against an already-resolved shared app.
-    async fn relay(
-        &self,
-        connection: &crate::connectors::GatewayConnection,
-        shared_app_id: &str,
-        body: &serde_json::Value,
-    ) -> std::result::Result<
-        crate::connectors::GatewayInvokeOutcome,
-        crate::connected_apps::GatewayDispatchError,
-    > {
-        use crate::connected_apps::GatewayDispatchError;
-
-        match connection.invoke_shared_app(shared_app_id, body).await {
-            Ok(Some(outcome)) => Ok(outcome),
-            Ok(None) => Err(GatewayDispatchError::NotRegistered),
-            // A session the gateway no longer honors is "no session" here, the
-            // same reading every other gateway read gives it.
-            Err(error) if is_sign_in_required(&error) => Err(GatewayDispatchError::NoSession),
-            Err(error) => {
-                tracing::warn!("gateway shared-app invoke failed: {error}");
-                Err(GatewayDispatchError::Unreachable(
-                    "the gateway could not complete this call".to_owned(),
-                ))
-            }
-        }
-    }
-}
-
-#[async_trait]
-impl crate::connected_apps::GatewayInvokeDispatcher for GatewayRelayDispatcher {
-    async fn dispatch(
-        &self,
-        owner: &tidebreak_core::OwnerId,
-        app: tidebreak_core::id::AppId,
-        request: &crate::connected_apps::GatewayOperationRequest,
-    ) -> std::result::Result<
-        crate::connectors::GatewayInvokeOutcome,
-        crate::connected_apps::GatewayDispatchError,
-    > {
-        use crate::connected_apps::GatewayDispatchError;
-
-        // A read that faults is reported as unreachable rather than as "no
-        // session": the distinction the route draws is whether a session
-        // exists, and a failed policy or vault read does not answer that.
-        let unreachable = |context: &'static str| {
-            move |error: tidebreak_core::AgentError| {
-                tracing::warn!("gateway relay could not {context}: {error}");
-                GatewayDispatchError::Unreachable(format!(
-                    "the gateway could not be reached to {context}"
-                ))
-            }
-        };
-        let policy = self
-            .runtime
-            .policy()
-            .map_err(unreachable("read this profile's gateway policy"))?;
-        let Some(base_url) = policy.gateway_url.clone().filter(|_| policy.managed) else {
-            return Err(GatewayDispatchError::NoSession);
-        };
-        let Some(connection) = self
-            .runtime
-            .connection_for(&policy)
-            .await
-            .map_err(unreachable("open a gateway connection"))?
-        else {
-            return Err(GatewayDispatchError::NoSession);
-        };
-        if connection
-            .stored_credentials()
-            .await
-            .map_err(unreachable("read the stored gateway session"))?
-            .is_none()
-        {
-            return Err(GatewayDispatchError::NoSession);
-        }
-        let shared_app_id = match self
-            .drafts
-            .ensure_registered(owner, app, &base_url)
-            .await
-            .map_err(unreachable("register this app at the gateway"))?
-        {
-            crate::connected_apps::GatewayRegistration::Registered { shared_app_id, .. } => {
-                shared_app_id
-            }
-            crate::connected_apps::GatewayRegistration::NotRegistered => {
-                return Err(GatewayDispatchError::NotRegistered)
-            }
-            crate::connected_apps::GatewayRegistration::Refused { message } => {
-                return Err(GatewayDispatchError::Unreachable(message))
-            }
-        };
-        let body = shared_app_invoke_body(request);
-        crate::gateway_drafts::relay_with_consent_self_heal(
-            &*self.drafts,
-            owner,
-            app,
-            &base_url,
-            || self.relay(&connection, &shared_app_id, &body),
-        )
-        .await
-    }
-}
-
-/// The gateway's own invoke vocabulary, verbatim. Absent halves are omitted
-/// rather than sent as null: the gateway's argument fields default when
-/// missing but refuse an explicit null, matching its `proxy_api` tool's
-/// schema — a null here makes every relayed call an `invalid_request`.
-fn shared_app_invoke_body(
-    request: &crate::connected_apps::GatewayOperationRequest,
-) -> serde_json::Value {
-    let mut body = serde_json::Map::new();
-    body.insert(
-        "connected_app_id".into(),
-        serde_json::Value::String(request.gateway_app.clone()),
-    );
-    body.insert(
-        "operation_id".into(),
-        serde_json::Value::String(request.operation_id.clone()),
-    );
-    if let Some(path_parameters) = &request.path_parameters {
-        body.insert("path_parameters".into(), path_parameters.clone());
-    }
-    if let Some(query) = &request.query {
-        body.insert("query".into(), query.clone());
-    }
-    if let Some(request_body) = &request.body {
-        body.insert("body".into(), request_body.clone());
-    }
-    serde_json::Value::Object(body)
-}
-
-/// Retire a stored gateway session the resolved policy no longer stands
-/// behind.
-///
-/// Two ways a session ends up superseded. An unmanaged profile can carry one
-/// signed in under the retired additive mode: nothing reaches it any more —
-/// the whole sign-in surface is managed-only, and the renderer has no
-/// gateway page. And a managed profile's policy authority can re-point the
-/// deployment (an MDM push): the old deployment's session is filtered out
-/// of every route and token path, but its refresh token stays live at the
-/// old gateway. Either way no surface will ever revoke it, so boot owns
-/// that cleanup.
-///
-/// Revocation is best-effort and bounded: an unreachable gateway can no more
-/// hold this hostage than it can a normal sign-out (the server-side session
-/// still dies at refresh-token expiry), and boot must not stall on it. The
-/// local clear afterwards is unconditional, so the session is gone locally
-/// whether or not the gateway ever answered — and because it is gone, this
-/// whole step runs at most once per superseded session.
-///
-/// An unreadable stored blob is left alone: it carries no usable refresh
-/// token, so it is not the live zombie this exists to kill. So is the
-/// session under a managed policy with no usable URL — that is a
-/// misconfigured policy to repair, not a supersession, and the session may
-/// well match it once repaired.
-pub(crate) async fn retire_superseded_gateway_session(
-    secrets: Arc<dyn SecretProvider>,
-    policy: &crate::managed_policy::ManagedPolicy,
-) -> Result<()> {
-    /// Long enough for a healthy gateway to answer a revoke, short enough
-    /// that a dead one is a hiccup at boot rather than a hang.
-    const REVOKE_TIMEOUT: Duration = Duration::from_secs(5);
-
-    let vault = CredentialVault::new(secrets.clone());
-    let Ok(Some(credentials)) = vault.load().await else {
-        return Ok(());
-    };
-    if policy.managed {
-        let Some(gateway_url) = policy.gateway_url.as_deref() else {
-            return Ok(());
-        };
-        if credentials.matches_base_url(gateway_url) {
-            return Ok(());
-        }
-        tracing::warn!(
-            "retiring the model-gateway session for {}: the managed policy \
-             now resolves {gateway_url}; sign in there to connect",
-            credentials.base_url
-        );
-    } else {
-        tracing::warn!(
-            "clearing a model-gateway session left by the retired additive \
-             configuration ({}); pair via your gateway's page to sign in again",
-            credentials.base_url
-        );
-    }
-    // The connection owns revoke-then-clear (the refresh token never leaves
-    // the connectors crate), so the session is retired through the same path
-    // an explicit sign-out takes.
-    if let Ok(config) = GatewayAuthConfig::new(&credentials.base_url) {
-        if let Ok(auth) = GatewayAuth::new(config) {
-            let connection = GatewayConnection::new(auth, CredentialVault::new(secrets.clone()));
-            let _ = tokio::time::timeout(REVOKE_TIMEOUT, connection.sign_out()).await;
-        }
-    }
-    // Unconditional: a gateway that never answered, or a stored base URL that
-    // no longer parses, must not leave the credential behind.
-    CredentialVault::new(secrets).clear().await
-}
-
-/// Compare the durable parts of two credentials while ignoring only cached
-/// access tokens. The refresh token is intentionally included: a replacement
-/// session for the same user and installation must still invalidate a catalog
-/// response authorized by the old session. A concurrent token rotation may
-/// conservatively reject a sync; the next background tick retries safely.
-fn same_gateway_session(
-    left: &crate::connectors::GatewayCredentials,
-    right: &crate::connectors::GatewayCredentials,
-) -> bool {
-    fn durable(credentials: &crate::connectors::GatewayCredentials) -> Option<serde_json::Value> {
-        let mut value = serde_json::to_value(credentials).ok()?;
-        value.as_object_mut()?.remove("access_tokens");
-        Some(value)
-    }
-
-    durable(left) == durable(right)
-}
-
-/// The one refusal for every managed-only gateway surface: unmanaged
-/// profiles have no gateway (policy is the only source), and a managed
-/// policy without a usable URL is misconfigured rather than open.
-fn require_managed(policy: &crate::managed_policy::ManagedPolicy) -> Result<String> {
-    if !policy.managed {
-        return Err(AgentError::config(
-            "this profile is not connected to a model gateway; \
-             pair via your gateway's page to connect",
-        ));
-    }
-    policy.gateway_url.clone().ok_or_else(|| {
-        AgentError::config(
-            "the managed gateway policy has no usable gateway URL; repair the policy authority",
-        )
-    })
-}
-
-/// Router-facing supplier of the gateway's `llm`-resource token. Refresh and
-/// rotation live inside [`GatewayConnection`]; this is just the seam.
-struct GatewayTokenSource {
-    connection: Arc<GatewayConnection>,
-    installation_id: String,
-    store: Arc<dyn Store>,
-    provisioned_policy: Arc<dyn crate::managed_policy::ProvisionedPolicySource>,
-    os_policy: Arc<dyn crate::managed_policy::OsPolicySource>,
-    model_sync: Arc<RwLock<()>>,
-}
-
-impl GatewayTokenSource {
-    /// Resolve one frozen selector against live policy/session/catalog state.
-    /// Callers decide which authority guard they retain; this helper never
-    /// acquires `model_sync`, so post-mint validation cannot recursively take
-    /// a fair read lock while a writer waits.
-    async fn resolve_model_route(
-        &self,
-        route_model: &str,
-    ) -> Result<Option<providers::ResolvedModelPolicy>> {
-        let policy = crate::managed_policy::resolve(&*self.provisioned_policy, &*self.os_policy)?;
-        let Some(snapshot) = providers::gateway_snapshot_for_policy(&*self.store, &policy).await?
-        else {
-            return Ok(None);
-        };
-        if snapshot.installation_id.as_deref() != Some(&self.installation_id) {
-            return Ok(None);
-        }
-        let Some(credentials) = self.connection.stored_credentials().await? else {
-            return Ok(None);
-        };
-        if credentials.installation_id != self.installation_id {
-            return Ok(None);
-        }
-        let selection = crate::model_registry::selection_key(
-            providers::ProviderKind::ModelGateway,
-            route_model,
-        );
-        let Some(resolved) = providers::gateway_execution_policy(&snapshot, &selection) else {
-            return Ok(None);
-        };
-        Ok((resolved.route_model == route_model).then_some(resolved))
-    }
-
-    fn route_lease(
-        resolved: providers::ResolvedModelPolicy,
-        guard: tokio::sync::OwnedRwLockReadGuard<()>,
-    ) -> ModelRouteLease {
-        ModelRouteLease::with_request_shaping_model(
-            resolved.id,
-            resolved.request_shaping_model,
-            guard,
-        )
-    }
-}
-
-#[async_trait]
-impl BearerTokenSource for GatewayTokenSource {
-    fn binding_id(&self) -> Option<&str> {
-        Some(&self.installation_id)
-    }
-
-    fn requires_model_route_lease(&self) -> bool {
-        true
-    }
-
-    async fn lease_model_route(&self, route_model: &str) -> Result<Option<ModelRouteLease>> {
-        let guard = self.model_sync.clone().read_owned().await;
-        Ok(self
-            .resolve_model_route(route_model)
-            .await?
-            .map(|resolved| Self::route_lease(resolved, guard)))
-    }
-
-    async fn authorize_model_route(
-        &self,
-        route_model: &str,
-        conversation: Option<tidebreak_core::id::ChatId>,
-    ) -> Result<(String, Option<ModelRouteLease>)> {
-        let guard = self.model_sync.clone().read_owned().await;
-        if self.resolve_model_route(route_model).await?.is_none() {
-            return Ok((String::new(), None));
-        }
-        let token = self.bearer_token_for(conversation).await?;
-        // OS/MDM policy does not participate in the process-local lock and may
-        // repoint while the network mint is delayed. Re-read every live input
-        // after that slow operation, using the guard already held so local
-        // sign-out/re-pair/deprovision remains fenced without recursive read
-        // acquisition.
-        let lease = self
-            .resolve_model_route(route_model)
-            .await?
-            .map(|resolved| Self::route_lease(resolved, guard));
-        Ok((token, lease))
-    }
-
-    async fn bearer_token(&self) -> Result<String> {
-        self.connection
-            .access_token_for_installation(RESOURCE_LLM, &self.installation_id)
-            .await
-    }
-
-    /// A chat's inference rides a token minted inside that chat's
-    /// attestation context, so the gateway records its tool calls as
-    /// observations and attested MCP endpoints can match them. Requests
-    /// with no conversation — titling, judging, other maintenance — keep
-    /// the shared token: there is no chat for an observation to serve.
-    async fn bearer_token_for(
-        &self,
-        conversation: Option<tidebreak_core::id::ChatId>,
-    ) -> Result<String> {
-        match conversation {
-            Some(chat) => {
-                self.connection
-                    .attested_access_token_for_installation(
-                        RESOURCE_LLM,
-                        &chat.to_string(),
-                        &self.installation_id,
-                    )
-                    .await
-            }
-            None => self.bearer_token().await,
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::collections::HashMap;
-    use std::sync::atomic::{AtomicUsize, Ordering};
-
-    use axum::extract::{Form, State};
-    use axum::http::{HeaderMap, Request};
-    use axum::response::{IntoResponse, Json, Response};
-    use axum::routing::{get, post};
-    use axum::{Extension, Router as AxumRouter};
-    use futures::StreamExt;
-    use serde_json::{json, Value};
-    use tidebreak_core::id::{AppId, AppRevisionId};
-    use tidebreak_core::local_app::{
-        AppGatewayOperationsGrantBinding, AppGrant, AppGrantBinding, AppManifest, CreateApp,
-        NewAppRevision,
-    };
-    use tidebreak_core::{
-        AgentConfig, ChatMessage, ChatRequest, Config, DbStore, ModelProvider, OwnerId, ProviderId,
-        Role, ToolRegistry,
-    };
-    use tower::ServiceExt;
-
-    use super::*;
-
-    /// The relay body is the gateway's `proxy_api` vocabulary: absent halves
-    /// must be absent keys, never null — the gateway's `serde(default)`
-    /// argument maps refuse an explicit null, so a regression here fails
-    /// every relayed call while passing any test that fakes the dispatcher.
-    #[test]
-    fn invoke_body_omits_absent_argument_halves() {
-        let bare = shared_app_invoke_body(&crate::connected_apps::GatewayOperationRequest {
-            gateway_app: "app-1".into(),
-            operation_id: "listMonitors".into(),
-            path_parameters: None,
-            query: None,
-            body: None,
-        });
-        assert_eq!(
-            bare,
-            json!({"connected_app_id": "app-1", "operation_id": "listMonitors"})
-        );
-
-        let full = shared_app_invoke_body(&crate::connected_apps::GatewayOperationRequest {
-            gateway_app: "app-1".into(),
-            operation_id: "getMonitor".into(),
-            path_parameters: Some(json!({"monitor_id": "7"})),
-            query: Some(json!({"verbose": true})),
-            body: Some(json!({"note": "hi"})),
-        });
-        assert_eq!(
-            full,
-            json!({
-                "connected_app_id": "app-1",
-                "operation_id": "getMonitor",
-                "path_parameters": {"monitor_id": "7"},
-                "query": {"verbose": true},
-                "body": {"note": "hi"},
-            })
-        );
-    }
-
-    #[derive(Default)]
-    struct MockSecrets(std::sync::Mutex<HashMap<String, String>>);
-
-    #[async_trait]
-    impl SecretProvider for MockSecrets {
-        async fn get_secret(&self, key: &str) -> Result<Option<String>> {
-            Ok(self.0.lock().unwrap().get(key).cloned())
-        }
-        async fn set_secret(&self, key: &str, value: &str) -> Result<()> {
-            self.0.lock().unwrap().insert(key.into(), value.into());
-            Ok(())
-        }
-        async fn delete_secret(&self, key: &str) -> Result<()> {
-            self.0.lock().unwrap().remove(key);
-            Ok(())
-        }
-    }
-
-    #[derive(Default)]
-    struct FakeGateway {
-        refreshes: AtomicUsize,
-        /// Number of metadata reads completed. Tests can delay the machine
-        /// offer until a later read to model a rolling Gateway deployment.
-        meta_reads: AtomicUsize,
-        machine_offer_after_reads: AtomicUsize,
-        /// Every refresh token posted to `/oauth/revoke`, in order.
-        revoked: std::sync::Mutex<Vec<String>>,
-        /// When set, `/api/v1/cli/apps` answers 500 — the outage shape the
-        /// endpoint-mount reconcile must degrade quietly on.
-        apps_fail: std::sync::atomic::AtomicBool,
-        /// `(resource, attestation_context_id)` per refresh grant, in order.
-        minted: std::sync::Mutex<Vec<(String, Option<String>)>>,
-        /// `(method, authorization)` per MCP endpoint request, in order.
-        mcp_requests: std::sync::Mutex<Vec<(String, String)>>,
-        /// When set, `/api/v1/me/catalog` serves this body under a fixed
-        /// `ETag`; when `None` the route answers 404, the shape of a gateway
-        /// that predates the member catalog.
-        catalog: std::sync::Mutex<Option<Value>>,
-        /// How many catalog fetches arrived with a matching `If-None-Match`
-        /// and were answered 304.
-        catalog_not_modified: AtomicUsize,
-    }
-
-    struct RepointableOs(std::sync::Mutex<String>);
-
-    impl RepointableOs {
-        fn new(base_url: String) -> Self {
-            Self(std::sync::Mutex::new(base_url))
-        }
-
-        fn repoint(&self, base_url: &str) {
-            *self.0.lock().unwrap() = base_url.to_owned();
-        }
-    }
-
-    impl crate::managed_policy::OsPolicySource for RepointableOs {
-        fn gateway_url(&self) -> Result<Option<String>> {
-            Ok(Some(self.0.lock().unwrap().clone()))
-        }
-    }
-
-    struct DelayedInferenceGateway {
-        delay_mint: usize,
-        mint_arrived: tokio::sync::Notify,
-        release_mint: tokio::sync::Notify,
-        llm_mints: AtomicUsize,
-        anthropic_requests: AtomicUsize,
-        responses_requests: AtomicUsize,
-        pause_first_anthropic_request: bool,
-    }
-
-    impl DelayedInferenceGateway {
-        fn new(delay_mint: usize, pause_first_anthropic_request: bool) -> Self {
-            Self {
-                delay_mint,
-                mint_arrived: tokio::sync::Notify::new(),
-                release_mint: tokio::sync::Notify::new(),
-                llm_mints: AtomicUsize::new(0),
-                anthropic_requests: AtomicUsize::new(0),
-                responses_requests: AtomicUsize::new(0),
-                pause_first_anthropic_request,
-            }
-        }
-    }
-
-    async fn delayed_inference_token(
-        State(gateway): State<Arc<DelayedInferenceGateway>>,
-        Form(form): Form<HashMap<String, String>>,
-    ) -> Json<Value> {
-        assert_eq!(
-            form.get("grant_type").map(String::as_str),
-            Some("refresh_token")
-        );
-        let resource = form.get("resource").cloned().unwrap_or_default();
-        let mint = gateway.llm_mints.fetch_add(1, Ordering::SeqCst);
-        if mint == gateway.delay_mint {
-            gateway.mint_arrived.notify_one();
-            gateway.release_mint.notified().await;
-        }
-        Json(json!({
-            "access_token": format!("mg_at_{resource}_{mint}"),
-            "token_type": "Bearer",
-            // Below the refresh leeway so every continuation performs the
-            // network-backed mint that the race is about.
-            "expires_in": 1,
-            "refresh_token": format!("mg_rt_{mint}"),
-            "scope": "inference:invoke",
-            "resource": resource,
-            "installation_id": "install-1",
-        }))
-    }
-
-    async fn delayed_anthropic_inference(
-        State(gateway): State<Arc<DelayedInferenceGateway>>,
-    ) -> Response {
-        let request = gateway.anthropic_requests.fetch_add(1, Ordering::SeqCst);
-        let stop_reason = if gateway.pause_first_anthropic_request && request == 0 {
-            "pause_turn"
-        } else {
-            "end_turn"
-        };
-        (
-            [(axum::http::header::CONTENT_TYPE, "text/event-stream")],
-            format!(
-                "data: {{\"type\":\"message_delta\",\"delta\":{{\"stop_reason\":\"{stop_reason}\"}}}}\n\n"
-            ),
-        )
-            .into_response()
-    }
-
-    async fn delayed_responses_inference(
-        State(gateway): State<Arc<DelayedInferenceGateway>>,
-    ) -> Response {
-        gateway.responses_requests.fetch_add(1, Ordering::SeqCst);
-        (
-            [(axum::http::header::CONTENT_TYPE, "text/event-stream")],
-            "data: {\"type\":\"response.completed\",\"response\":{}}\n\n",
-        )
-            .into_response()
-    }
-
-    async fn managed_router_with_delayed_inference(
-        protocol: providers::GatewayModelProtocol,
-        gateway: Arc<DelayedInferenceGateway>,
-    ) -> (
-        tidebreak_router::Router,
-        String,
-        Arc<RepointableOs>,
-        tempfile::TempDir,
-    ) {
-        let app = AxumRouter::new()
-            .route("/oauth/token", post(delayed_inference_token))
-            .route(
-                "/compat/anthropic/v1/messages",
-                post(delayed_anthropic_inference),
-            )
-            .route(
-                "/compat/openai/v1/responses",
-                post(delayed_responses_inference),
-            )
-            .with_state(gateway);
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let base = format!("http://{}", listener.local_addr().unwrap());
-        tokio::spawn(async move { axum::serve(listener, app).await.unwrap() });
-
-        let directory = tempfile::tempdir().unwrap();
-        let store: Arc<dyn Store> = Arc::new(
-            DbStore::connect(&format!(
-                "sqlite://{}?mode=rwc",
-                directory.path().join("gateway.db").display()
-            ))
-            .await
-            .unwrap(),
-        );
-        let secrets: Arc<dyn SecretProvider> = Arc::new(MockSecrets::default());
-        let credentials: crate::connectors::GatewayCredentials = serde_json::from_value(json!({
-            "base_url": base,
-            "installation_id": "install-1",
-            "user_id": "user-1",
-            "refresh_token": "mg_rt_seed",
-            "access_tokens": {}
-        }))
-        .unwrap();
-        CredentialVault::new(secrets.clone())
-            .save(&credentials)
-            .await
-            .unwrap();
-        let os = Arc::new(RepointableOs::new(base.clone()));
-        let runtime = GatewayRuntime::new(
-            store.clone(),
-            secrets,
-            crate::managed_policy::MemoryProvisionedPolicy::new(),
-            os.clone(),
-        );
-        let policy = runtime.policy().unwrap();
-        let (model, upstream) = match protocol {
-            providers::GatewayModelProtocol::AnthropicMessages => {
-                ("anthropic-us-claude-opus-5", "us.anthropic.claude-opus-5")
-            }
-            providers::GatewayModelProtocol::OpenaiResponses => {
-                ("gateway-gpt-5-6-sol", "gpt-5.6-sol")
-            }
-        };
-        let snapshot = providers::GatewayModelSnapshot {
-            gateway_url: policy.gateway_url.clone().unwrap(),
-            installation_id: Some("install-1".into()),
-            models: vec![CustomModelConfig {
-                id: model.into(),
-                upstream_id: Some(upstream.into()),
-                ..Default::default()
-            }],
-            model_protocols: std::collections::BTreeMap::from([(model.into(), protocol)]),
-            model_reasoning_efforts: Default::default(),
-            member_catalog: Some("v1".into()),
-            catalog_etag: None,
-        };
-        providers::write_gateway_snapshot(&*store, &snapshot)
-            .await
-            .unwrap();
-        let frozen = providers::gateway_execution_policy(
-            &snapshot,
-            &crate::model_registry::selection_key(providers::ProviderKind::ModelGateway, model),
-        )
-        .unwrap()
-        .route_model;
-        let routes = providers::collect_routes(
-            &*store,
-            &*runtime.secrets,
-            runtime.route_token_source().await,
-            None,
-            None,
-            &policy,
-        )
-        .await;
-        (
-            tidebreak_router::Router::build(routes),
-            frozen,
-            os,
-            directory,
-        )
-    }
-
-    #[tokio::test]
-    async fn external_policy_repoint_during_bearer_mint_blocks_gateway_initial_egress() {
-        for protocol in [
-            providers::GatewayModelProtocol::AnthropicMessages,
-            providers::GatewayModelProtocol::OpenaiResponses,
-        ] {
-            let gateway = Arc::new(DelayedInferenceGateway::new(0, false));
-            let (router, frozen, os, _directory) =
-                managed_router_with_delayed_inference(protocol, gateway.clone()).await;
-            let request = ChatRequest {
-                provider: Some(ProviderId::new("model_gateway")),
-                model: frozen,
-                messages: vec![ChatMessage::text(Role::User, "hi")],
-                ..Default::default()
-            };
-            let dispatch = tokio::spawn(async move { router.stream(request).await });
-
-            tokio::time::timeout(Duration::from_secs(2), gateway.mint_arrived.notified())
-                .await
-                .expect("the inference token mint reaches gateway A");
-            os.repoint("https://gateway-b.example.test");
-            gateway.release_mint.notify_one();
-
-            let result = tokio::time::timeout(Duration::from_secs(2), dispatch)
-                .await
-                .expect("authorization finishes after the delayed mint")
-                .unwrap();
-            assert!(
-                result.is_err(),
-                "a token minted by a gateway retired during refresh must not authorize egress"
-            );
-            assert_eq!(
-                gateway.anthropic_requests.load(Ordering::SeqCst),
-                0,
-                "Anthropic request data must not reach retired gateway A"
-            );
-            assert_eq!(
-                gateway.responses_requests.load(Ordering::SeqCst),
-                0,
-                "Responses request data must not reach retired gateway A"
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn external_policy_repoint_during_continuation_mint_blocks_anthropic_egress() {
-        let gateway = Arc::new(DelayedInferenceGateway::new(1, true));
-        let (router, frozen, os, _directory) = managed_router_with_delayed_inference(
-            providers::GatewayModelProtocol::AnthropicMessages,
-            gateway.clone(),
-        )
-        .await;
-        let stream = router
-            .stream(ChatRequest {
-                provider: Some(ProviderId::new("model_gateway")),
-                model: frozen,
-                vendor_web_search: Some(tidebreak_core::provider::VendorWebSearch { max_uses: 1 }),
-                messages: vec![ChatMessage::text(Role::User, "search")],
-                ..Default::default()
-            })
-            .await
-            .unwrap();
-        assert_eq!(gateway.anthropic_requests.load(Ordering::SeqCst), 1);
-
-        let continuation = tokio::spawn(async move { stream.collect::<Vec<_>>().await });
-        tokio::time::timeout(Duration::from_secs(2), gateway.mint_arrived.notified())
-            .await
-            .expect("the continuation starts its fresh token mint");
-        os.repoint("https://gateway-b.example.test");
-        gateway.release_mint.notify_one();
-
-        let events = tokio::time::timeout(Duration::from_secs(2), continuation)
-            .await
-            .expect("the continuation refuses the retired route")
-            .unwrap();
-        assert_eq!(
-            gateway.anthropic_requests.load(Ordering::SeqCst),
-            1,
-            "only the leg dispatched before the MDM repoint may reach gateway A"
-        );
-        assert_eq!(gateway.llm_mints.load(Ordering::SeqCst), 2);
-        assert!(matches!(
-            events.last(),
-            Some(tidebreak_core::ProviderEvent::Failed { .. })
-        ));
-    }
-
-    const CATALOG_ETAG: &str = "W/\"catalog-rev-1\"";
-
-    async fn catalog(State(gateway): State<Arc<FakeGateway>>, headers: HeaderMap) -> Response {
-        let Some(body) = gateway.catalog.lock().unwrap().clone() else {
-            return axum::http::StatusCode::NOT_FOUND.into_response();
-        };
-        let bearer = headers
-            .get("authorization")
-            .and_then(|value| value.to_str().ok())
-            .unwrap_or_default();
-        assert!(bearer.starts_with("Bearer mg_at_control_"), "{bearer}");
-        if headers
-            .get("if-none-match")
-            .and_then(|value| value.to_str().ok())
-            == Some(CATALOG_ETAG)
-        {
-            gateway.catalog_not_modified.fetch_add(1, Ordering::SeqCst);
-            return (
-                axum::http::StatusCode::NOT_MODIFIED,
-                [(axum::http::header::ETAG, CATALOG_ETAG)],
-            )
-                .into_response();
-        }
-        ([(axum::http::header::ETAG, CATALOG_ETAG)], Json(body)).into_response()
-    }
-
-    /// One entitled connected app aggregating the fixture's `tools` MCP
-    /// endpoint, the shape `/api/v1/cli/apps` serves.
-    async fn apps(State(gateway): State<Arc<FakeGateway>>, headers: HeaderMap) -> Response {
-        if gateway.apps_fail.load(Ordering::SeqCst) {
-            return (
-                axum::http::StatusCode::INTERNAL_SERVER_ERROR,
-                "apps are down",
-            )
-                .into_response();
-        }
-        let bearer = headers
-            .get("authorization")
-            .and_then(|value| value.to_str().ok())
-            .unwrap_or_default();
-        assert!(bearer.starts_with("Bearer mg_at_control_"), "{bearer}");
-        Json(json!({
-            "apps": [{
-                "id": "app-1",
-                "name": "Tools",
-                "app_kind": "mcp_endpoint",
-                "enabled": true,
-                "mcp_endpoint_slugs": ["tools"]
-            }]
-        }))
-        .into_response()
-    }
-
-    async fn token(
-        State(gateway): State<Arc<FakeGateway>>,
-        Form(form): Form<HashMap<String, String>>,
-    ) -> Json<Value> {
-        assert_eq!(
-            form.get("grant_type").map(String::as_str),
-            Some("refresh_token")
-        );
-        let sequence = gateway.refreshes.fetch_add(1, Ordering::SeqCst);
-        let resource = form.get("resource").cloned().unwrap_or_default();
-        gateway.minted.lock().unwrap().push((
-            resource.clone(),
-            form.get("attestation_context_id").cloned(),
-        ));
-        Json(json!({
-            "access_token": format!("mg_at_{resource}_{sequence}"),
-            "token_type": "Bearer",
-            "expires_in": 600,
-            "refresh_token": format!("mg_rt_{sequence}"),
-            "scope": "models:read inference:invoke",
-            "resource": resource,
-            "installation_id": "install-1",
-        }))
-    }
-
-    async fn models(
-        axum::extract::Query(query): axum::extract::Query<HashMap<String, String>>,
-        headers: HeaderMap,
-    ) -> Response {
-        assert!(!query.contains_key("protocol"));
-        let bearer = headers
-            .get("authorization")
-            .and_then(|value| value.to_str().ok())
-            .unwrap_or_default();
-        assert!(bearer.starts_with("Bearer mg_at_control_"), "{bearer}");
-        Json(json!({
-            "models": [
-                {
-                    "id": "sample-claude",
-                    "protocol": "anthropic_messages",
-                    "name": "Sample Claude",
-                    "context_window": 200000,
-                    "max_output_tokens": 8192,
-                    "supports_tools": true,
-                    "supports_vision": true
-                },
-                {
-                    "id": "sample-coder",
-                    "protocol": "openai_responses",
-                    "name": "Sample Coder",
-                    "context_window": null,
-                    "max_output_tokens": null,
-                    "supports_tools": true,
-                    "supports_vision": false
-                }
-            ]
-        }))
-        .into_response()
-    }
-
-    /// Minimal Streamable HTTP MCP endpoint that requires an `mcp:tools`
-    /// session bearer, exactly as the gateway's `/mcp/{slug}` route does.
-    async fn mcp_endpoint(
-        State(gateway): State<Arc<FakeGateway>>,
-        headers: HeaderMap,
-        body: String,
-    ) -> Response {
-        let bearer = headers
-            .get("authorization")
-            .and_then(|value| value.to_str().ok())
-            .unwrap_or_default();
-        assert!(bearer.starts_with("Bearer mg_at_mcp:tools_"), "{bearer}");
-        let request: Value = serde_json::from_str(&body).unwrap();
-        let method = request["method"].as_str().unwrap_or_default().to_string();
-        gateway
-            .mcp_requests
-            .lock()
-            .unwrap()
-            .push((method.clone(), bearer.to_string()));
-        let id = request.get("id").cloned().unwrap_or_default();
-        let result = match method.as_str() {
-            "initialize" => json!({
-                "protocolVersion": tidebreak_mcp::PROTOCOL_VERSION,
-                "capabilities": {"tools": {}},
-                "serverInfo": {"name": "gateway-fixture", "version": "1"}
-            }),
-            "tools/list" => json!({
-                "tools": [{
-                    "name": "lookup",
-                    "description": "Look something up",
-                    "inputSchema": {"type": "object"}
-                }]
-            }),
-            "tools/call" => json!({
-                "content": [{"type": "text", "text": "ok"}]
-            }),
-            _ => json!({}),
-        };
-        Json(json!({"jsonrpc": "2.0", "id": id, "result": result})).into_response()
-    }
-
-    async fn revoke(
-        State(gateway): State<Arc<FakeGateway>>,
-        Form(form): Form<HashMap<String, String>>,
-    ) -> Json<Value> {
-        gateway
-            .revoked
-            .lock()
-            .unwrap()
-            .push(form.get("token").cloned().unwrap_or_default());
-        Json(json!({}))
-    }
-
-    async fn meta(State(gateway): State<Arc<FakeGateway>>) -> Json<Value> {
-        let read = gateway.meta_reads.fetch_add(1, Ordering::SeqCst);
-        let offer_after = gateway.machine_offer_after_reads.load(Ordering::SeqCst);
-        let mut metadata = json!({
-            "api_version": "v1",
-            "installation_id": "install-1",
-            "gateway_version": "1.0.0",
-            "public_url": "http://gateway.test",
-            "auth_mode": "oidc",
-        });
-        if read >= offer_after {
-            metadata["tidebreak_machine_url"] = json!("https://machine.tidebreak.test");
-        }
-        Json(metadata)
-    }
-
-    async fn serve(gateway: Arc<FakeGateway>) -> std::net::SocketAddr {
-        let app = AxumRouter::new()
-            .route("/api/v1/meta", get(meta))
-            .route("/oauth/token", post(token))
-            .route("/oauth/revoke", post(revoke))
-            .route("/api/v1/cli/models", get(models))
-            .route("/api/v1/cli/apps", get(apps))
-            .route("/api/v1/me/catalog", get(catalog))
-            .route("/mcp/{slug}", post(mcp_endpoint))
-            .with_state(gateway);
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let address = listener.local_addr().unwrap();
-        tokio::spawn(async move {
-            axum::serve(listener, app).await.unwrap();
-        });
-        address
-    }
-
-    /// A runtime with a stored session for `session_base`, on a profile
-    /// provisioned (managed) to `provisioned` — policy being the only way a
-    /// profile is gateway-connected at all.
-    async fn signed_in_runtime_at(
-        session_base: &str,
-        provisioned: &str,
-    ) -> (Arc<GatewayRuntime>, Arc<dyn Store>, tempfile::TempDir) {
-        let directory = tempfile::tempdir().unwrap();
-        let store: Arc<dyn Store> = Arc::new(
-            DbStore::connect(&format!(
-                "sqlite://{}?mode=rwc",
-                directory.path().join("gateway.db").display()
-            ))
-            .await
-            .unwrap(),
-        );
-        let secrets: Arc<dyn SecretProvider> = Arc::new(MockSecrets::default());
-        // Stored credentials are private to the connectors crate by design;
-        // seed the vault through its serialized form, exactly as a completed
-        // sign-in would have persisted it.
-        let credentials: crate::connectors::GatewayCredentials = serde_json::from_value(json!({
-            "base_url": session_base,
-            "installation_id": "install-1",
-            "user_id": "user-1",
-            "account_hint": "abaas@example.test",
-            "refresh_token": "mg_rt_seed",
-            "access_tokens": {}
-        }))
-        .unwrap();
-        CredentialVault::new(secrets.clone())
-            .save(&credentials)
-            .await
-            .unwrap();
-        let provisioned_policy = crate::managed_policy::MemoryProvisionedPolicy::new();
-        crate::managed_policy::provision(&*provisioned_policy, provisioned).unwrap();
-        (
-            GatewayRuntime::new(
-                store.clone(),
-                secrets,
-                provisioned_policy,
-                Arc::new(crate::managed_policy::NoOsPolicy),
-            ),
-            store,
-            directory,
-        )
-    }
-
-    async fn signed_in_runtime(
-        base_url: &str,
-    ) -> (Arc<GatewayRuntime>, Arc<dyn Store>, tempfile::TempDir) {
-        signed_in_runtime_at(base_url, base_url).await
-    }
-
-    /// An MCP runtime resolving gateway endpoints through `runtime`, the way
-    /// the server wires the two together.
-    fn mcp_for(
-        runtime: &Arc<GatewayRuntime>,
-        store: &Arc<dyn Store>,
-    ) -> Arc<crate::mcp_config::McpRuntime> {
-        Arc::new(crate::mcp_config::McpRuntime::new(
-            Arc::new(tidebreak_core::ToolRegistry::new()),
-            store.clone(),
-            Arc::new(MockSecrets::default()),
-            runtime.clone(),
-            runtime.provisioned_policy().clone(),
-            Arc::new(crate::managed_policy::NoOsPolicy),
-        ))
-    }
-
-    /// A gateway that hosts a machine names it on `/api/v1/meta`, and the
-    /// settings panel prefills the address field from it. Reading it needs
-    /// no session: meta is unauthenticated, so the offer is there before the
-    /// first sign-in.
-    #[tokio::test]
-    async fn the_offered_machine_comes_from_gateway_meta() {
-        let address = serve(Arc::new(FakeGateway::default())).await;
-        let base = format!("http://{address}");
-        let (runtime, _store, _directory) = signed_in_runtime(&base).await;
-
-        assert_eq!(
-            runtime.offered_machine().await.url.as_deref(),
-            Some("https://machine.tidebreak.test")
-        );
-    }
-
-    #[tokio::test]
-    async fn a_missing_machine_offer_is_retried_during_the_same_process() {
-        let gateway = Arc::new(FakeGateway {
-            machine_offer_after_reads: AtomicUsize::new(1),
-            ..Default::default()
-        });
-        let address = serve(gateway.clone()).await;
-        let base = format!("http://{address}");
-        let (runtime, _store, _directory) = signed_in_runtime(&base).await;
-
-        assert!(runtime.offered_machine().await.url.is_none());
-        assert_eq!(
-            runtime.offered_machine().await.url.as_deref(),
-            Some("https://machine.tidebreak.test")
-        );
-        assert_eq!(
-            runtime.offered_machine().await.url.as_deref(),
-            Some("https://machine.tidebreak.test")
-        );
-        assert_eq!(gateway.meta_reads.load(Ordering::SeqCst), 2);
-    }
-
-    #[tokio::test]
-    async fn syncs_entitled_models_into_the_snapshot() {
-        let address = serve(Arc::new(FakeGateway::default())).await;
-        let base = format!("http://{address}");
-        let (runtime, store, _directory) = signed_in_runtime(&base).await;
-
-        assert_eq!(runtime.sync_models().await.unwrap(), 2);
-
-        let policy = runtime.policy().unwrap();
-        let models = providers::gateway_models(&*store, &policy, None)
-            .await
-            .unwrap();
-        assert_eq!(models.len(), 2);
-        assert_eq!(models[0].id, "sample-claude");
-        assert_eq!(models[0].display_name.as_deref(), Some("Sample Claude"));
-        assert_eq!(models[0].context_window, 200_000);
-        // Absent limits fall back to the conservative custom-model defaults.
-        assert_eq!(models[1].context_window, 32_768);
-        assert_eq!(models[1].max_output_tokens, 4_096);
-        let snapshot = providers::read_gateway_snapshot(&*store)
-            .await
-            .unwrap()
-            .expect("the sync persisted its protocol map");
-        assert_eq!(
-            snapshot.model_protocols.get("sample-claude"),
-            Some(&providers::GatewayModelProtocol::AnthropicMessages)
-        );
-        assert_eq!(
-            snapshot.model_protocols.get("sample-coder"),
-            Some(&providers::GatewayModelProtocol::OpenaiResponses)
-        );
-
-        // The synced snapshot resolves as a model policy under the gateway key.
-        let policy =
-            providers::resolve_model_policy(&*store, "model_gateway::sample-claude", false, None)
-                .await
-                .unwrap()
-                .expect("synced model resolves");
-        assert_eq!(
-            policy.provider,
-            crate::providers::ProviderKind::ModelGateway
-        );
-        assert_eq!(policy.display_name, "Sample Claude");
-        // Synced through the degraded CLI reads: no catalog revision to stamp.
-        let snapshot = providers::read_gateway_snapshot(&*store).await.unwrap();
-        assert_eq!(snapshot.unwrap().member_catalog, None);
-    }
-
-    fn sample_catalog() -> Value {
-        json!({
-            "models": [
-                {
-                    "id": "sample-claude",
-                    "name": "Sample Claude",
-                    // Dual-protocol: one row, both surfaces, and the sync
-                    // must route it through Anthropic Messages.
-                    "protocols": ["anthropic_messages", "openai_responses"],
-                    "aliases": ["us.anthropic.claude-opus-5"],
-                    "supports_tools": true,
-                    "supports_vision": true,
-                    "context_window": 200000,
-                    "max_output_tokens": 8192,
-                    "provider_name": "Anthropic"
-                },
-                {
-                    "id": "sample-coder",
-                    "name": "Sample Coder",
-                    "protocols": ["openai_responses"],
-                    "supports_tools": true,
-                    "supports_vision": false,
-                    "context_window": null,
-                    "max_output_tokens": null,
-                    "provider_name": "Acme"
-                },
-                {
-                    "id": "sample-exotic",
-                    "name": "Sample Exotic",
-                    // A protocol this client does not speak: skipped, never
-                    // fatal — the set is the gateway's to grow.
-                    "protocols": ["grpc_frames"],
-                    "supports_tools": false,
-                    "supports_vision": false,
-                    "context_window": null,
-                    "max_output_tokens": null,
-                    "provider_name": "Acme"
-                }
-            ],
-            "apps": [{
-                "id": "app-1",
-                "name": "Tools",
-                "app_kind": "mcp_endpoint",
-                "enabled": true,
-                "mcp_endpoint_slugs": ["tools"],
-                "connection": "authorization_required"
-            }]
-        })
-    }
-
-    #[tokio::test]
-    async fn sync_prefers_the_member_catalog_and_stamps_its_revision() {
-        let gateway = Arc::new(FakeGateway::default());
-        *gateway.catalog.lock().unwrap() = Some(sample_catalog());
-        let address = serve(gateway).await;
-        let base = format!("http://{address}");
-        let (runtime, store, _directory) = signed_in_runtime(&base).await;
-
-        // Two models the client can route; the exotic protocol is skipped.
-        assert_eq!(runtime.sync_models().await.unwrap(), 2);
-
-        let snapshot = providers::read_gateway_snapshot(&*store)
-            .await
-            .unwrap()
-            .expect("the sync persisted a snapshot");
-        assert_eq!(snapshot.member_catalog.as_deref(), Some("v1"));
-        assert_eq!(snapshot.catalog_etag.as_deref(), Some(CATALOG_ETAG));
-        assert_eq!(
-            snapshot.model_protocols.get("sample-claude"),
-            Some(&providers::GatewayModelProtocol::AnthropicMessages),
-            "a dual-protocol model routes through Anthropic Messages"
-        );
-        assert_eq!(
-            snapshot.model_protocols.get("sample-coder"),
-            Some(&providers::GatewayModelProtocol::OpenaiResponses)
-        );
-        assert!(!snapshot.model_protocols.contains_key("sample-exotic"));
-
-        // The catalog's alias matches the curated registry row, so the
-        // policy carries curated capabilities under the gateway's own id.
-        let policy =
-            providers::resolve_model_policy(&*store, "model_gateway::sample-claude", false, None)
-                .await
-                .unwrap()
-                .expect("synced model resolves");
-        assert_eq!(policy.display_name, "Claude Opus 5");
-
-        // The status projection surfaces the catalog revision for the
-        // settings panel's older-gateway note.
-        let status = runtime.status().await.unwrap();
-        assert_eq!(status.member_catalog.as_deref(), Some("v1"));
-    }
-
-    #[tokio::test]
-    async fn overlapping_model_syncs_serialize_before_the_second_fetch() {
-        let gateway = Arc::new(FakeGateway::default());
-        let requests = Arc::new(AtomicUsize::new(0));
-        let (arrived_tx, mut arrived_rx) = tokio::sync::mpsc::unbounded_channel::<usize>();
-        let catalog_route = {
-            let requests = requests.clone();
-            move |headers: HeaderMap| {
-                let requests = requests.clone();
-                let arrived = arrived_tx.clone();
-                async move {
-                    let bearer = headers
-                        .get("authorization")
-                        .and_then(|value| value.to_str().ok())
-                        .unwrap_or_default();
-                    assert!(bearer.starts_with("Bearer mg_at_control_"), "{bearer}");
-                    let request = requests.fetch_add(1, Ordering::SeqCst);
-                    arrived.send(request).expect("the test is listening");
-                    (
-                        [(axum::http::header::ETAG, format!("W/\"rev-{request}\""))],
-                        Json(sample_catalog()),
-                    )
-                        .into_response()
-                }
-            }
-        };
-        let app = AxumRouter::new()
-            .route("/oauth/token", post(token))
-            .route("/api/v1/me/catalog", get(catalog_route))
-            .with_state(gateway);
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let base = format!("http://{}", listener.local_addr().unwrap());
-        tokio::spawn(async move {
-            axum::serve(listener, app).await.unwrap();
-        });
-        let (runtime, store, _directory) = signed_in_runtime(&base).await;
-        let first_pause = Arc::new(MigrationPause::default());
-        *runtime.sync_commit_pause.lock().await = Some(first_pause.clone());
-
-        let first = tokio::spawn({
-            let runtime = runtime.clone();
-            async move { runtime.sync_models().await }
-        });
-        assert_eq!(arrived_rx.recv().await, Some(0));
-        first_pause.arrived.notified().await;
-
-        let (second_started_tx, second_started_rx) = tokio::sync::oneshot::channel();
-        let second = tokio::spawn({
-            let runtime = runtime.clone();
-            async move {
-                second_started_tx.send(()).unwrap();
-                runtime.sync_models().await
-            }
-        });
-        second_started_rx.await.unwrap();
-        tokio::task::yield_now().await;
-        assert_eq!(
-            requests.load(Ordering::SeqCst),
-            1,
-            "the second sync must wait before issuing its catalog fetch"
-        );
-
-        first_pause.release.notify_one();
-        assert_eq!(first.await.unwrap().unwrap(), 2);
-        assert_eq!(arrived_rx.recv().await, Some(1));
-        assert_eq!(second.await.unwrap().unwrap(), 2);
-        assert_eq!(requests.load(Ordering::SeqCst), 2);
-        assert_eq!(
-            providers::read_gateway_snapshot(&*store)
-                .await
-                .unwrap()
-                .unwrap()
-                .catalog_etag
-                .as_deref(),
-            Some("W/\"rev-1\"")
-        );
-    }
-
-    #[tokio::test]
-    async fn request_route_setups_share_read_leases_while_catalog_sync_waits() {
-        let gateway = Arc::new(FakeGateway::default());
-        *gateway.catalog.lock().unwrap() = Some(sample_catalog());
-        let address = serve(gateway.clone()).await;
-        let base = format!("http://{address}");
-        let (runtime, store, _directory) = signed_in_runtime(&base).await;
-        runtime.sync_models().await.unwrap();
-
-        let policy = runtime.policy().unwrap();
-        let snapshot = providers::gateway_snapshot_for_policy(&*store, &policy)
-            .await
-            .unwrap()
-            .unwrap();
-        let frozen = providers::gateway_execution_policy(&snapshot, "model_gateway::sample-claude")
-            .unwrap()
-            .route_model;
-        let source = runtime.route_token_source().await.unwrap();
-
-        let first = source
-            .lease_model_route(&frozen)
-            .await
-            .unwrap()
-            .expect("first request setup receives a route lease");
-        let second =
-            tokio::time::timeout(Duration::from_secs(1), source.lease_model_route(&frozen))
-                .await
-                .expect("a second request setup must not serialize behind the first")
-                .unwrap()
-                .expect("second request setup receives the same live route lease");
-
-        let mut writer = tokio::spawn({
-            let runtime = runtime.clone();
-            async move { runtime.sync_models().await }
-        });
-        assert!(
-            tokio::time::timeout(Duration::from_millis(50), &mut writer)
-                .await
-                .is_err(),
-            "catalog mutation must wait while request legs hold read leases"
-        );
-        assert_eq!(gateway.catalog_not_modified.load(Ordering::SeqCst), 0);
-
-        drop(first);
-        assert!(
-            tokio::time::timeout(Duration::from_millis(50), &mut writer)
-                .await
-                .is_err(),
-            "every active request leg must release before the writer proceeds"
-        );
-        drop(second);
-        assert_eq!(
-            tokio::time::timeout(Duration::from_secs(2), writer)
-                .await
-                .expect("catalog sync proceeds after both dispatch leases end")
-                .unwrap()
-                .unwrap(),
-            2
-        );
-        assert_eq!(gateway.catalog_not_modified.load(Ordering::SeqCst), 1);
-    }
-
-    #[tokio::test]
-    async fn sign_out_waits_for_catalog_sync_then_clears_its_commit() {
-        let gateway = Arc::new(FakeGateway::default());
-        *gateway.catalog.lock().unwrap() = Some(sample_catalog());
-        let address = serve(gateway).await;
-        let base = format!("http://{address}");
-        let (runtime, store, _directory) = signed_in_runtime(&base).await;
-        crate::model_roles::write_selection(
-            &*store,
-            crate::model_roles::ModelRole::Chat,
-            Some("anthropic::claude-opus-5"),
-        )
-        .await
-        .unwrap();
-        let pause = Arc::new(MigrationPause::default());
-        *runtime.sync_commit_pause.lock().await = Some(pause.clone());
-
-        let sync = tokio::spawn({
-            let runtime = runtime.clone();
-            async move { runtime.sync_models().await }
-        });
-        pause.arrived.notified().await;
-        let mut sign_out = tokio::spawn({
-            let runtime = runtime.clone();
-            async move { runtime.sign_out().await }
-        });
-        assert!(
-            tokio::time::timeout(Duration::from_millis(50), &mut sign_out)
-                .await
-                .is_err(),
-            "sign-out must wait for the in-flight catalog writer instead of deadlocking it"
-        );
-        pause.release.notify_one();
-        assert_eq!(sync.await.unwrap().unwrap(), 2);
-        tokio::time::timeout(Duration::from_secs(2), sign_out)
-            .await
-            .expect("sign-out proceeds after catalog sync releases the writer")
-            .unwrap()
-            .unwrap();
-
-        let policy = runtime.policy().unwrap();
-        assert!(providers::gateway_models(&*store, &policy, None)
-            .await
-            .unwrap()
-            .is_empty());
-        assert!(runtime
-            .connection()
-            .await
-            .unwrap()
-            .unwrap()
-            .stored_credentials()
-            .await
-            .unwrap()
-            .is_none());
-        assert_eq!(
-            crate::model_roles::read_selection(&*store, crate::model_roles::ModelRole::Chat)
-                .await
-                .unwrap()
-                .as_deref(),
-            Some("anthropic::claude-opus-5")
-        );
-    }
-
-    #[tokio::test]
-    async fn a_catalog_response_cannot_commit_after_same_url_session_replacement() {
-        let gateway = Arc::new(FakeGateway::default());
-        *gateway.catalog.lock().unwrap() = Some(sample_catalog());
-        let address = serve(gateway).await;
-        let base = format!("http://{address}");
-        let (runtime, store, _directory) = signed_in_runtime(&base).await;
-        crate::model_roles::write_selection(
-            &*store,
-            crate::model_roles::ModelRole::Chat,
-            Some("anthropic::claude-opus-5"),
-        )
-        .await
-        .unwrap();
-        let pause = Arc::new(MigrationPause::default());
-        *runtime.sync_commit_pause.lock().await = Some(pause.clone());
-
-        let sync = tokio::spawn({
-            let runtime = runtime.clone();
-            async move { runtime.sync_models().await }
-        });
-        pause.arrived.notified().await;
-        let replacement: crate::connectors::GatewayCredentials = serde_json::from_value(json!({
-            "base_url": base,
-            "installation_id": "install-1",
-            "user_id": "user-1",
-            "account_hint": "abaas@example.test",
-            "refresh_token": "mg_rt_replacement",
-            "access_tokens": {}
-        }))
-        .unwrap();
-        CredentialVault::new(runtime.secrets.clone())
-            .save(&replacement)
-            .await
-            .unwrap();
-        pause.release.notify_one();
-
-        let error = sync
-            .await
-            .unwrap()
-            .expect_err("the old session's response must not commit over its replacement");
-        assert_eq!(error.kind(), "gateway_changed");
-        assert!(providers::read_gateway_snapshot(&*store)
-            .await
-            .unwrap()
-            .is_none());
-        assert_eq!(
-            crate::model_roles::read_selection(&*store, crate::model_roles::ModelRole::Chat)
-                .await
-                .unwrap()
-                .as_deref(),
-            Some("anthropic::claude-opus-5")
-        );
-    }
-
-    #[tokio::test]
-    async fn sync_never_rewrites_durable_model_selections() {
-        let gateway = Arc::new(FakeGateway::default());
-        *gateway.catalog.lock().unwrap() = Some(sample_catalog());
-        let address = serve(gateway).await;
-        let base = format!("http://{address}");
-        let (runtime, store, _directory) = signed_in_runtime(&base).await;
-
-        crate::model_roles::write_selection(
-            &*store,
-            crate::model_roles::ModelRole::Chat,
-            Some("anthropic::claude-opus-5"),
-        )
-        .await
-        .unwrap();
-        store
-            .set_setting(
-                "chat_default.model",
-                &serde_json::Value::String("claude-opus-5".into()),
-            )
-            .await
-            .unwrap();
-        let pinned = tidebreak_core::Chat {
-            id: tidebreak_core::ChatId::new(),
-            project_id: None,
-            title: None,
-            model: Some("claude-opus-5".into()),
-            reasoning_effort: None,
-            permission_mode: None,
-            network_policy: Default::default(),
-            attachment_revision: 0,
-            root_attachments: Vec::new(),
-            created_at: chrono::Utc::now(),
-        };
-        store.create_chat(&pinned).await.unwrap();
-
-        runtime.sync_models().await.unwrap();
-
-        assert_eq!(
-            crate::model_roles::read_selection(&*store, crate::model_roles::ModelRole::Chat)
-                .await
-                .unwrap()
-                .as_deref(),
-            Some("anthropic::claude-opus-5")
-        );
-        assert_eq!(
-            store.get_setting("chat_default.model").await.unwrap(),
-            Some(serde_json::Value::String("claude-opus-5".into()))
-        );
-        assert_eq!(
-            store
-                .get_chat(pinned.id)
-                .await
-                .unwrap()
-                .unwrap()
-                .model
-                .as_deref(),
-            Some("claude-opus-5")
-        );
-    }
-
-    #[tokio::test]
-    async fn an_unchanged_catalog_answers_304_and_the_snapshot_stays() {
-        let gateway = Arc::new(FakeGateway::default());
-        *gateway.catalog.lock().unwrap() = Some(sample_catalog());
-        let address = serve(gateway.clone()).await;
-        let base = format!("http://{address}");
-        let (runtime, store, _directory) = signed_in_runtime(&base).await;
-
-        assert_eq!(runtime.sync_models().await.unwrap(), 2);
-        // The second tick sends the stored ETag and keeps the snapshot.
-        assert_eq!(runtime.sync_models().await.unwrap(), 2);
-        assert_eq!(gateway.catalog_not_modified.load(Ordering::SeqCst), 1);
-        let snapshot = providers::read_gateway_snapshot(&*store).await.unwrap();
-        assert_eq!(snapshot.unwrap().models.len(), 2);
-    }
-
-    #[tokio::test]
-    async fn a_304_sync_racing_a_policy_repoint_is_rejected() {
-        let gateway = Arc::new(FakeGateway::default());
-        *gateway.catalog.lock().unwrap() = Some(sample_catalog());
-        let address = serve(gateway.clone()).await;
-        let base = format!("http://{address}");
-
-        struct SwappableOs(std::sync::Mutex<String>);
-
-        impl crate::managed_policy::OsPolicySource for SwappableOs {
-            fn gateway_url(&self) -> Result<Option<String>> {
-                Ok(Some(self.0.lock().unwrap().clone()))
-            }
-        }
-
-        let (seed_runtime, store, directory) = signed_in_runtime(&base).await;
-        seed_runtime.sync_models().await.unwrap();
-        drop(seed_runtime);
-        let os = Arc::new(SwappableOs(std::sync::Mutex::new(base.clone())));
-        let secrets: Arc<dyn SecretProvider> = Arc::new(MockSecrets::default());
-        let credentials: crate::connectors::GatewayCredentials = serde_json::from_value(json!({
-            "base_url": base,
-            "installation_id": "install-1",
-            "user_id": "user-1",
-            "refresh_token": "mg_rt_seed",
-            "access_tokens": {}
-        }))
-        .unwrap();
-        CredentialVault::new(secrets.clone())
-            .save(&credentials)
-            .await
-            .unwrap();
-        let runtime = GatewayRuntime::new(
-            store.clone(),
-            secrets,
-            crate::managed_policy::MemoryProvisionedPolicy::new(),
-            os.clone(),
-        );
-        let held = providers::GATEWAY_STATE_WRITES.lock().await;
-        let sync = tokio::spawn({
-            let runtime = runtime.clone();
-            async move { runtime.sync_models().await }
-        });
-        while gateway.catalog_not_modified.load(Ordering::SeqCst) == 0 {
-            tokio::task::yield_now().await;
-        }
-        *os.0.lock().unwrap() = "https://gateway-b.test".to_owned();
-        drop(held);
-
-        let error = sync
-            .await
-            .unwrap()
-            .expect_err("a 304 sync must recheck policy before returning");
-        assert_eq!(error.kind(), "gateway_changed");
-        drop(directory);
-    }
-
-    #[tokio::test]
-    async fn apps_carry_readiness_from_the_member_catalog() {
-        let gateway = Arc::new(FakeGateway::default());
-        *gateway.catalog.lock().unwrap() = Some(sample_catalog());
-        let address = serve(gateway).await;
-        let base = format!("http://{address}");
-        let (runtime, _store, _directory) = signed_in_runtime(&base).await;
-
-        let apps = runtime
-            .apps(&tidebreak_core::OwnerId::local())
-            .await
-            .unwrap();
-        assert!(apps.supported);
-        assert_eq!(apps.apps.len(), 1);
-        assert_eq!(
-            apps.apps[0].connection.as_deref(),
-            Some("authorization_required")
-        );
-    }
-
-    #[tokio::test]
-    async fn gateway_apps_route_scopes_used_by_count_to_the_authenticated_principal() {
-        let gateway = Arc::new(FakeGateway::default());
-        *gateway.catalog.lock().unwrap() = Some(sample_catalog());
-        let address = serve(gateway).await;
-        let base = format!("http://{address}");
-        let (runtime, store, directory) = signed_in_runtime(&base).await;
-        let alice = OwnerId::new("user:alice").unwrap();
-        let app_id = AppId::new();
-        store
-            .create_app_scoped(
-                &alice,
-                &CreateApp {
-                    id: app_id,
-                    revision: NewAppRevision {
-                        id: AppRevisionId::new(),
-                        manifest: AppManifest {
-                            name: "Alice gateway app".into(),
-                            bindings: Vec::new(),
-                        },
-                        byte_len: 1,
-                        sha256: [0; 32],
-                        turn_id: None,
-                        producing_run_id: None,
-                        chat_id: None,
-                        created_at: chrono::Utc::now(),
-                    },
-                },
-            )
-            .await
-            .unwrap();
-        store
-            .put_app_grant_scoped(
-                &alice,
-                &AppGrant {
-                    app_id,
-                    bindings: vec![AppGrantBinding::GatewayOperations(
-                        AppGatewayOperationsGrantBinding {
-                            gateway_app: "app-1".into(),
-                            operation_ids: vec!["list".into()],
-                            fingerprint: [1; 32],
-                        },
-                    )],
-                    created_at: chrono::Utc::now(),
-                },
-            )
-            .await
-            .unwrap();
-
-        struct RouteTestResolver;
-
-        #[async_trait]
-        impl crate::resolver::ProviderResolver for RouteTestResolver {
-            async fn resolve(&self) -> Arc<dyn ModelProvider> {
-                Arc::new(crate::provider::UnconfiguredProvider)
-            }
-        }
-
-        let secrets: Arc<dyn SecretProvider> = Arc::new(MockSecrets::default());
-        let chatgpt = Arc::new(
-            crate::chatgpt_runtime::ChatGptRuntime::new(store.clone(), secrets.clone()).unwrap(),
-        );
-        let state = crate::state::AppState::with_gateway_runtime(
-            Config::desktop(directory.path()),
-            store,
-            Arc::new(RouteTestResolver),
-            secrets,
-            Arc::new(ToolRegistry::new()),
-            AgentConfig::default(),
-            uuid::Uuid::new_v4(),
-            runtime,
-            chatgpt,
-            crate::managed_policy::MemoryProvisionedPolicy::new(),
-            Arc::new(crate::managed_policy::NoOsPolicy),
-        )
-        .unwrap();
-
-        async fn count_for(
-            state: crate::state::AppState,
-            principal: crate::principal::Principal,
-        ) -> usize {
-            let response = AxumRouter::new()
-                .route("/gateway/apps", get(crate::routes::get_gateway_apps))
-                .with_state(state)
-                .layer(Extension(crate::principal::AuthContext {
-                    principal,
-                    client_executor: false,
-                }))
-                .oneshot(
-                    Request::builder()
-                        .uri("/gateway/apps")
-                        .body(axum::body::Body::empty())
-                        .unwrap(),
-                )
-                .await
-                .unwrap();
-            assert_eq!(response.status(), axum::http::StatusCode::OK);
-            let body = axum::body::to_bytes(response.into_body(), usize::MAX)
-                .await
-                .unwrap();
-            serde_json::from_slice::<Value>(&body).unwrap()["apps"][0]["used_by_app_count"]
-                .as_u64()
-                .unwrap() as usize
-        }
-
-        let alice_count = count_for(
-            state.clone(),
-            crate::principal::Principal::User {
-                id: crate::principal::UserId::new("alice").unwrap(),
-                role: crate::principal::Role::Member,
-            },
-        )
-        .await;
-        let bob_count = count_for(
-            state,
-            crate::principal::Principal::User {
-                id: crate::principal::UserId::new("bob").unwrap(),
-                role: crate::principal::Role::Member,
-            },
-        )
-        .await;
-
-        assert_eq!(alice_count, 1);
-        assert_eq!(bob_count, 0);
-    }
-
-    #[tokio::test]
-    async fn apps_from_a_catalogless_gateway_carry_no_readiness() {
-        let address = serve(Arc::new(FakeGateway::default())).await;
-        let base = format!("http://{address}");
-        let (runtime, _store, _directory) = signed_in_runtime(&base).await;
-
-        let apps = runtime
-            .apps(&tidebreak_core::OwnerId::local())
-            .await
-            .unwrap();
-        assert!(apps.supported);
-        assert_eq!(apps.apps.len(), 1);
-        assert_eq!(apps.apps[0].connection, None);
-    }
-
-    /// The boot-before-sign-in race: a status or sync read caches the
-    /// keychain's `NoEntry`, and the session then lands without the cache
-    /// observing the write — a read that was in flight while sign-in
-    /// completed stores its stale miss *after* the write's invalidation, and
-    /// a session written by another process never invalidates at all. The
-    /// gateway session key's misses are not memoized
-    /// (`CachingSecretProvider::with_miss_passthrough`), so the next
-    /// background tick sees the session with no restart and no manual sync.
-    #[tokio::test]
-    async fn a_session_written_behind_the_secret_cache_syncs_without_a_restart() {
-        let address = serve(Arc::new(FakeGateway::default())).await;
-        let base = format!("http://{address}");
-        let directory = tempfile::tempdir().unwrap();
-        let store: Arc<dyn Store> = Arc::new(
-            DbStore::connect(&format!(
-                "sqlite://{}?mode=rwc",
-                directory.path().join("gateway.db").display()
-            ))
-            .await
-            .unwrap(),
-        );
-        let inner = Arc::new(MockSecrets::default());
-        // The same wrap `secret_provider` in lib.rs applies at boot.
-        let secrets: Arc<dyn SecretProvider> = Arc::new(
-            tidebreak_core::CachingSecretProvider::new(inner.clone())
-                .with_miss_passthrough([crate::connectors::GATEWAY_SECRET_KEY]),
-        );
-        let provisioned_policy = crate::managed_policy::MemoryProvisionedPolicy::new();
-        crate::managed_policy::provision(&*provisioned_policy, &base).unwrap();
-        let runtime = GatewayRuntime::new(
-            store.clone(),
-            secrets,
-            provisioned_policy,
-            Arc::new(crate::managed_policy::NoOsPolicy),
-        );
-
-        // Boot before sign-in: the tick reads the empty store. Were the miss
-        // memoized, this is the read that would hide the session forever.
-        assert_eq!(runtime.sync_models_if_connected().await.unwrap(), None);
-
-        // The session lands behind the cache's back, straight into the store.
-        let credentials: crate::connectors::GatewayCredentials = serde_json::from_value(json!({
-            "base_url": base,
-            "installation_id": "install-1",
-            "user_id": "user-1",
-            "account_hint": "abaas@example.test",
-            "refresh_token": "mg_rt_seed",
-            "access_tokens": {}
-        }))
-        .unwrap();
-        inner
-            .set_secret(
-                crate::connectors::GATEWAY_SECRET_KEY,
-                &serde_json::to_string(&credentials).unwrap(),
-            )
-            .await
-            .unwrap();
-
-        // The very next tick recovers — no restart, no manual sync click.
-        assert_eq!(runtime.sync_models_if_connected().await.unwrap(), Some(2));
-    }
-
-    /// A gateway that serves a model under its upstream id is serving that
-    /// curated model over a different route, so the row inherits the curated
-    /// capabilities instead of being presented as an anonymous unverified
-    /// endpoint — whether the gateway id is itself the curated id or a
-    /// deployment alias whose reported upstream id resolves to one. The
-    /// deployment's own limits still win, and a row the registry can be matched
-    /// to by neither route keeps the conservative treatment.
-    #[tokio::test]
-    async fn a_gateway_model_matching_a_curated_id_inherits_its_capabilities() {
-        let address = serve(Arc::new(FakeGateway::default())).await;
-        let base = format!("http://{address}");
-        let (_runtime, store, _directory) = signed_in_runtime(&base).await;
-
-        providers::write_gateway_snapshot(
-            &*store,
-            &providers::GatewayModelSnapshot {
-                gateway_url: base.clone(),
-                installation_id: Some("install-1".into()),
-                models: vec![
-                    CustomModelConfig {
-                        id: "claude-opus-5".to_string(),
-                        upstream_id: None,
-                        display_name: Some("Opus (gateway)".to_string()),
-                        context_window: 200_000,
-                        max_output_tokens: 32_000,
-                        ..Default::default()
-                    },
-                    CustomModelConfig {
-                        id: "anthropic-us-claude-opus-5".to_string(),
-                        upstream_id: Some("us.anthropic.claude-opus-5".to_string()),
-                        display_name: None,
-                        context_window: 200_000,
-                        max_output_tokens: 32_000,
-                        ..Default::default()
-                    },
-                    CustomModelConfig {
-                        id: "acme-inhouse-llm".to_string(),
-                        upstream_id: None,
-                        display_name: None,
-                        context_window: 200_000,
-                        max_output_tokens: 32_000,
-                        ..Default::default()
-                    },
-                ],
-                model_protocols: Default::default(),
-                model_reasoning_efforts: Default::default(),
-                member_catalog: None,
-                catalog_etag: None,
-            },
-        )
-        .await
-        .unwrap();
-
-        let matched =
-            providers::resolve_model_policy(&*store, "model_gateway::claude-opus-5", false, None)
-                .await
-                .unwrap()
-                .expect("the gateway row resolves");
-        assert_eq!(
-            matched.provider,
-            crate::providers::ProviderKind::ModelGateway,
-            "the gateway still serves the request"
-        );
-        assert_eq!(
-            matched.vendor,
-            Some(crate::providers::ProviderKind::Anthropic)
-        );
-        assert_eq!(matched.key, "model_gateway::claude-opus-5");
-        assert_eq!(matched.display_name, "Claude Opus 5");
-        assert_eq!(
-            matched.verification,
-            crate::model_registry::VerificationTier::Verified
-        );
-        assert!(matched
-            .input_modalities
-            .contains(&crate::model_registry::InputModality::Image));
-        assert!(matched.supports_reasoning);
-        assert!(matched.supports_structured_output);
-        assert!(!matched.reasoning_efforts.is_empty());
-        // The deployment's reported limits, not the curated model's.
-        assert_eq!(matched.context_window, 200_000);
-        assert_eq!(matched.max_output_tokens, 32_000);
-
-        let aliased = providers::resolve_model_policy(
-            &*store,
-            "model_gateway::anthropic-us-claude-opus-5",
-            false,
-            None,
-        )
-        .await
-        .unwrap()
-        .expect("the aliased gateway row resolves");
-        assert_eq!(
-            aliased.provider,
-            crate::providers::ProviderKind::ModelGateway,
-            "matching by upstream id must not re-route the request"
-        );
-        assert_eq!(
-            aliased.id, "anthropic-us-claude-opus-5",
-            "the gateway's own id is what the request carries"
-        );
-        assert_eq!(
-            aliased.request_shaping_model, "claude-opus-5",
-            "the unambiguous canonical upstream identity shapes the request"
-        );
-        assert_eq!(
-            aliased.vendor,
-            Some(crate::providers::ProviderKind::Anthropic)
-        );
-        assert_eq!(aliased.display_name, "Claude Opus 5");
-        assert_eq!(
-            aliased.verification,
-            crate::model_registry::VerificationTier::Verified
-        );
-        assert!(aliased
-            .input_modalities
-            .contains(&crate::model_registry::InputModality::Image));
-        assert!(aliased.supports_reasoning);
-        assert!(aliased.supports_structured_output);
-        assert_eq!(aliased.context_window, 200_000);
-
-        let unmatched = providers::resolve_model_policy(
-            &*store,
-            "model_gateway::acme-inhouse-llm",
-            false,
-            None,
-        )
-        .await
-        .unwrap()
-        .expect("the unmatched gateway row still resolves");
-        assert_eq!(unmatched.vendor, None);
-        assert_eq!(
-            unmatched.verification,
-            crate::model_registry::VerificationTier::Unverified
-        );
-        assert_eq!(
-            unmatched.input_modalities,
-            vec![crate::model_registry::InputModality::Text]
-        );
-        assert!(!unmatched.supports_reasoning);
-        assert!(!unmatched.supports_structured_output);
-    }
-
-    /// The background loop's first attempt is immediate — the boot case it
-    /// exists for: a signed-in profile gets a fresh snapshot without anyone
-    /// pressing Refresh.
-    #[tokio::test]
-    async fn the_background_sync_populates_the_snapshot_without_a_manual_refresh() {
-        let address = serve(Arc::new(FakeGateway::default())).await;
-        let base = format!("http://{address}");
-        let (runtime, store, _directory) = signed_in_runtime(&base).await;
-
-        let task = tokio::spawn(
-            runtime
-                .clone()
-                .sync_models_periodically(mcp_for(&runtime, &store)),
-        );
-        let synced = async {
-            loop {
-                if let Some(snapshot) = providers::read_gateway_snapshot(&*store).await.unwrap() {
-                    break snapshot;
-                }
-                tokio::time::sleep(Duration::from_millis(10)).await;
-            }
-        };
-        let snapshot = tokio::time::timeout(Duration::from_secs(5), synced)
-            .await
-            .expect("the boot-time sync lands without a manual refresh");
-        task.abort();
-        assert_eq!(snapshot.models.len(), 2);
-    }
-
-    #[tokio::test]
-    async fn the_route_token_source_mints_llm_tokens_per_rotation() {
-        let gateway = Arc::new(FakeGateway::default());
-        let address = serve(gateway.clone()).await;
-        let base = format!("http://{address}");
-        let (runtime, store, _directory) = signed_in_runtime(&base).await;
-
-        let source = runtime
-            .route_token_source()
-            .await
-            .expect("signed-in runtime offers a token source");
-        let token = source.bearer_token().await.unwrap();
-        assert!(token.starts_with("mg_at_llm_"), "{token}");
-        // A fresh token is cached: no second refresh inside the expiry leeway.
-        assert_eq!(source.bearer_token().await.unwrap(), token);
-        assert_eq!(gateway.refreshes.load(Ordering::SeqCst), 1);
-
-        // A conversation mints inside that chat's attestation context: cached
-        // per chat, distinct across chats, and never the shared token.
-        let chat = tidebreak_core::id::ChatId::new();
-        let attested = source.bearer_token_for(Some(chat)).await.unwrap();
-        assert_ne!(attested, token);
-        assert_eq!(source.bearer_token_for(Some(chat)).await.unwrap(), attested);
-        let other = source
-            .bearer_token_for(Some(tidebreak_core::id::ChatId::new()))
-            .await
-            .unwrap();
-        assert_ne!(other, attested);
-        // No conversation — titling, judging, maintenance — keeps the shared
-        // token and records nothing.
-        assert_eq!(source.bearer_token_for(None).await.unwrap(), token);
-
-        // The route set includes the gateway with its synced models claimed.
-        // A legacy provider row — even one pointing at a different deployment
-        // — is never read, so it changes nothing about the composite route.
-        runtime.sync_models().await.unwrap();
-        providers::write_config(
-            &*store,
-            crate::providers::ProviderKind::ModelGateway,
-            &providers::ProviderConfig {
-                enabled: true,
-                base_url: Some("http://127.0.0.1:9".to_string()),
-                models: Vec::new(),
-            },
-        )
-        .await
-        .unwrap();
-        let policy = runtime.policy().unwrap();
-        let routes = providers::collect_routes(
-            &*store,
-            &*runtime.secrets,
-            runtime.route_token_source().await,
-            None,
-            None,
-            &policy,
-        )
-        .await;
-        let anthropic_route = routes
-            .iter()
-            .find(|route| route.kind == tidebreak_router::RouteKind::ModelGateway)
-            .expect("Anthropic gateway route present");
-        assert_eq!(
-            anthropic_route.base_url.as_deref(),
-            Some(format!("{base}/compat/anthropic").as_str())
-        );
-        assert!(anthropic_route.api_key.is_empty());
-        assert!(anthropic_route
-            .curated_models
-            .contains(&"sample-claude".to_string()));
-        assert!(!anthropic_route
-            .curated_models
-            .contains(&"sample-coder".to_string()));
-        let anthropic_frozen: Vec<_> = anthropic_route
-            .curated_models
-            .iter()
-            .filter(|model| model.starts_with("__tidebreak_gateway_v1."))
-            .collect();
-        assert_eq!(anthropic_frozen.len(), 1);
-        assert_eq!(
-            anthropic_route.model_rewrites.get(anthropic_frozen[0]),
-            Some(&"sample-claude".to_string())
-        );
-        let openai_route = routes
-            .iter()
-            .find(|route| route.kind == tidebreak_router::RouteKind::ModelGatewayOpenai)
-            .expect("OpenAI gateway route present");
-        assert_eq!(
-            openai_route.base_url.as_deref(),
-            Some(format!("{base}/compat/openai/v1").as_str())
-        );
-        assert!(openai_route.api_key.is_empty());
-        assert!(openai_route
-            .curated_models
-            .contains(&"sample-coder".to_string()));
-        let openai_frozen: Vec<_> = openai_route
-            .curated_models
-            .iter()
-            .filter(|model| model.starts_with("__tidebreak_gateway_v1."))
-            .collect();
-        assert_eq!(openai_frozen.len(), 1);
-        assert_eq!(
-            openai_route.model_rewrites.get(openai_frozen[0]),
-            Some(&"sample-coder".to_string())
-        );
-        assert!(providers::provider_is_usable(
-            &*store,
-            &*runtime.secrets,
-            crate::providers::ProviderKind::ModelGateway,
-            &policy,
-            None
-        )
-        .await
-        .unwrap());
-
-        let frozen = anthropic_frozen[0].as_str();
-        let lease = source
-            .lease_model_route(frozen)
-            .await
-            .unwrap()
-            .expect("the current frozen selector receives a live route lease");
-        assert_eq!(lease.wire_model(), "sample-claude");
-        assert_eq!(lease.request_shaping_model(), "sample-claude");
-        drop(lease);
-
-        let mut retargeted = providers::gateway_snapshot_for_policy(&*store, &policy)
-            .await
-            .unwrap()
-            .unwrap();
-        retargeted
-            .models
-            .iter_mut()
-            .find(|model| model.id == "sample-claude")
-            .unwrap()
-            .upstream_id = Some("gpt-5.6-sol".into());
-        providers::write_gateway_snapshot(&*store, &retargeted)
-            .await
-            .unwrap();
-        assert!(source.lease_model_route(frozen).await.unwrap().is_none());
-    }
-
-    /// The race #896 named, in its surviving form: with policy the only
-    /// gateway source, the one authority that can re-point a deployment
-    /// mid-sync is an OS (MDM) push. A sync whose entitlement fetch is still
-    /// in flight when that happens must not stamp the old gateway's model
-    /// list into the snapshot — what this pins is the under-lock policy
-    /// recheck refusing the stale write.
-    #[tokio::test]
-    async fn a_sync_racing_a_policy_repoint_cannot_stamp_the_old_gateways_models() {
-        // Gateway A: answers token refreshes normally but parks the model
-        // list until released — the window the pairing lands in.
-        let (arrived_tx, mut arrived_rx) = tokio::sync::mpsc::unbounded_channel::<()>();
-        let release = Arc::new(tokio::sync::Notify::new());
-        let parked_models = {
-            let release = release.clone();
-            move || {
-                let release = release.clone();
-                let arrived = arrived_tx.clone();
-                async move {
-                    arrived.send(()).expect("the test is listening");
-                    release.notified().await;
-                    Json(json!({
-                        "models": [{
-                            "id": "stale-model",
-                            "name": "Stale Model",
-                            "context_window": 200000,
-                            "max_output_tokens": 8192,
-                            "supports_tools": true,
-                            "supports_vision": false
-                        }]
-                    }))
-                }
-            }
-        };
-        let app = AxumRouter::new()
-            .route("/oauth/token", post(token))
-            .route("/api/v1/cli/models", get(parked_models))
-            .with_state(Arc::new(FakeGateway::default()));
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let base_a = format!("http://{}", listener.local_addr().unwrap());
-        tokio::spawn(async move {
-            axum::serve(listener, app).await.unwrap();
-        });
-
-        // An OS authority whose asserted deployment can change mid-test, as
-        // an MDM push would.
-        struct SwappableOs(std::sync::Mutex<String>);
-
-        impl crate::managed_policy::OsPolicySource for SwappableOs {
-            fn gateway_url(&self) -> Result<Option<String>> {
-                Ok(Some(self.0.lock().unwrap().clone()))
-            }
-        }
-
-        let directory = tempfile::tempdir().unwrap();
-        let store: Arc<dyn Store> = Arc::new(
-            DbStore::connect(&format!(
-                "sqlite://{}?mode=rwc",
-                directory.path().join("gateway.db").display()
-            ))
-            .await
-            .unwrap(),
-        );
-        let secrets: Arc<dyn SecretProvider> = Arc::new(MockSecrets::default());
-        let credentials: crate::connectors::GatewayCredentials = serde_json::from_value(json!({
-            "base_url": base_a,
-            "installation_id": "install-1",
-            "user_id": "user-1",
-            "refresh_token": "mg_rt_seed",
-            "access_tokens": {}
-        }))
-        .unwrap();
-        CredentialVault::new(secrets.clone())
-            .save(&credentials)
-            .await
-            .unwrap();
-        let os = Arc::new(SwappableOs(std::sync::Mutex::new(base_a.clone())));
-        let runtime = GatewayRuntime::new(
-            store.clone(),
-            secrets,
-            crate::managed_policy::MemoryProvisionedPolicy::new(),
-            os.clone(),
-        );
-
-        let sync = tokio::spawn({
-            let runtime = runtime.clone();
-            async move { runtime.sync_models().await }
-        });
-        arrived_rx
-            .recv()
-            .await
-            .expect("the fetch reaches gateway A");
-
-        // While A's model list is in flight, the MDM authority re-points the
-        // profile at gateway B.
-        *os.0.lock().unwrap() = "https://gateway-b.test".to_string();
-
-        release.notify_one();
-        let error = sync
-            .await
-            .unwrap()
-            .expect_err("a sync whose deployment changed mid-fetch must refuse to write");
-        assert_eq!(
-            error.kind(),
-            "gateway_changed",
-            "the refusal is the stable retryable conflict, not a fault: {error:?}"
-        );
-
-        assert!(
-            providers::read_gateway_snapshot(&*store)
-                .await
-                .unwrap()
-                .is_none(),
-            "gateway A's models must not be stamped into the snapshot"
-        );
-    }
-
-    #[tokio::test]
-    async fn status_reflects_the_session_and_sign_out_clears_the_snapshot() {
-        let address = serve(Arc::new(FakeGateway::default())).await;
-        let base = format!("http://{address}");
-        let (runtime, store, _directory) = signed_in_runtime(&base).await;
-        runtime.sync_models().await.unwrap();
-
-        let status = runtime.status().await.unwrap();
-        assert!(status.base_url.is_some() && status.signed_in);
-        assert_eq!(status.account_hint.as_deref(), Some("abaas@example.test"));
-        assert_eq!(status.installation_id.as_deref(), Some("install-1"));
-        assert_eq!(status.model_count, 2);
-        assert_eq!(status.sign_in, SignInProgress::Idle);
-        // The projection carries no token-shaped material.
-        let json = serde_json::to_string(&status).unwrap();
-        assert!(!json.contains("mg_at_"));
-        assert!(!json.contains("mg_rt_"));
-
-        runtime.sign_out().await.unwrap();
-        let status = runtime.status().await.unwrap();
-        assert!(!status.signed_in);
-        assert_eq!(status.model_count, 0);
-        assert!(status.account_hint.is_none());
-        let policy = runtime.policy().unwrap();
-        assert!(providers::gateway_models(&*store, &policy, None)
-            .await
-            .unwrap()
-            .is_empty());
-        // Managed policy is a separate layer with a separate lifecycle:
-        // disconnecting the session must never deprovision the profile.
-        assert!(policy.managed, "sign-out must not deprovision the profile");
-    }
-
-    #[tokio::test]
-    async fn mounts_a_gateway_endpoint_from_the_signed_in_session() {
-        let gateway = Arc::new(FakeGateway::default());
-        let address = serve(gateway.clone()).await;
-        let base = format!("http://{address}");
-        let (runtime, store, _directory) = signed_in_runtime(&base).await;
-
-        let mcp = mcp_for(&runtime, &store);
-        let definition = crate::mcp_config::McpServerDefinition {
-            name: "tools".to_string(),
-            command: None,
-            args: Vec::new(),
-            env: std::collections::BTreeSet::new(),
-            env_values: std::collections::BTreeMap::new(),
-            env_from: Vec::new(),
-            cwd: None,
-            url: None,
-            bearer_token_env: None,
-            gateway_endpoint: Some("tools".to_string()),
-            request_timeout_ms: 60_000,
-            enabled: true,
-            plugin: None,
-            launch: None,
-        };
-        // No environment variable anywhere: the URL and bearer come from the
-        // stored session via a resource-scoped refresh (asserted inside the
-        // fixture endpoint).
-        let info = mcp
-            .replace(crate::mcp_config::McpServersConfig {
-                servers: vec![definition],
-            })
-            .await
-            .unwrap();
-        assert_eq!(
-            info.servers[0].health,
-            crate::mcp_config::McpHealth::Healthy
-        );
-        assert!(mcp.snapshot().get("mcp__tools__lookup").is_some());
-
-        // The handshake rode a context-bearing token: an attested endpoint
-        // refuses any request minted without one, so the connect mint is what
-        // makes attested mounts able to list their tools at all.
-        let connect_context = {
-            let minted = gateway.minted.lock().unwrap();
-            let (_, context) = minted
-                .iter()
-                .find(|(resource, _)| resource == "mcp:tools")
-                .expect("the mount minted an mcp:tools token")
-                .clone();
-            context.expect("the connect mint carries an attestation context")
-        };
-
-        // A tool call from a chat presents that chat's bearer, minted inside
-        // the same attestation context the chat's inference tokens carry —
-        // the invariant that lets an attested endpoint match the call against
-        // the observation the inference recorded.
-        let chat = tidebreak_core::id::ChatId::new();
-        let snapshot = mcp.snapshot();
-        let tool = snapshot.get("mcp__tools__lookup").unwrap();
-        let ctx = tidebreak_core::ToolCtx::without_private_scratch(chat, None);
-        tool.execute(&ctx, json!({})).await.unwrap();
-
-        let source = runtime
-            .route_token_source()
-            .await
-            .expect("signed-in runtime offers a token source");
-        source.bearer_token_for(Some(chat)).await.unwrap();
-        {
-            let minted = gateway.minted.lock().unwrap();
-            let chat_mcp = minted
-                .iter()
-                .filter(|(resource, _)| resource == "mcp:tools")
-                .nth(1)
-                .map(|(_, context)| context.clone().unwrap())
-                .expect("the tool call minted a second mcp:tools token");
-            let chat_llm = minted
-                .iter()
-                .find(|(resource, _)| resource == "llm")
-                .map(|(_, context)| context.clone().unwrap())
-                .expect("the chat's inference minted an llm token");
-            assert_ne!(chat_mcp, connect_context);
-            assert_eq!(chat_mcp, chat_llm);
-        }
-        {
-            let requests = gateway.mcp_requests.lock().unwrap();
-            let handshake: Vec<&str> = requests
-                .iter()
-                .filter(|(method, _)| method == "initialize" || method == "tools/list")
-                .map(|(_, bearer)| bearer.as_str())
-                .collect();
-            assert!(!handshake.is_empty());
-            let call = requests
-                .iter()
-                .find(|(method, _)| method == "tools/call")
-                .map(|(_, bearer)| bearer.as_str())
-                .expect("the tool call reached the endpoint");
-            assert!(handshake.iter().all(|bearer| *bearer == handshake[0]));
-            assert_ne!(call, handshake[0]);
-        }
-
-        // Sign-out degrades the mount to a secret-free sign-in diagnostic and
-        // keeps the definition; the tool leaves the registry.
-        runtime.sign_out().await.unwrap();
-        let error = mcp.reconnect("tools").await.err().unwrap();
-        assert!(!error.to_string().contains("mg_at_"), "{error}");
-        let info = mcp.info().await;
-        assert_eq!(
-            info.servers[0].health,
-            crate::mcp_config::McpHealth::Degraded
-        );
-        assert_eq!(
-            info.servers[0].diagnostic.as_deref(),
-            Some("Sign in to the model gateway to reconnect this server.")
-        );
-        assert_eq!(
-            info.servers[0].definition.gateway_endpoint.as_deref(),
-            Some("tools")
-        );
-        assert!(mcp.snapshot().get("mcp__tools__lookup").is_none());
-    }
-
-    /// #1441: the assembled `AppState` must give the resolver, the /gateway
-    /// routes, and MCP dispatch ONE gateway runtime. Attestation contexts
-    /// live in a per-instance registry, so when `state.mcp` held its own
-    /// runtime a chat's inference token and its MCP call bearer minted in
-    /// two different contexts and the gateway's observation-consume refused
-    /// every attested `tools/call`. Assembled the way `bind_inner` does —
-    /// one runtime injected into the resolver and the state — and driven
-    /// across the boundary that used to be two instances: the token source
-    /// from `state.gateway`, the dispatch bearer through `state.mcp`.
-    #[tokio::test]
-    async fn assembled_state_mints_inference_and_mcp_dispatch_in_one_attestation_context() {
-        let gateway = Arc::new(FakeGateway::default());
-        let address = serve(gateway.clone()).await;
-        let base = format!("http://{address}");
-        let directory = tempfile::tempdir().unwrap();
-        let store: Arc<dyn Store> = Arc::new(
-            DbStore::connect(&format!(
-                "sqlite://{}?mode=rwc",
-                directory.path().join("gateway.db").display()
-            ))
-            .await
-            .unwrap(),
-        );
-        let secrets: Arc<dyn SecretProvider> = Arc::new(MockSecrets::default());
-        let credentials: crate::connectors::GatewayCredentials = serde_json::from_value(json!({
-            "base_url": base,
-            "installation_id": "install-1",
-            "user_id": "user-1",
-            "refresh_token": "mg_rt_seed",
-            "access_tokens": {}
-        }))
-        .unwrap();
-        CredentialVault::new(secrets.clone())
-            .save(&credentials)
-            .await
-            .unwrap();
-        let provisioned_policy = crate::managed_policy::MemoryProvisionedPolicy::new();
-        crate::managed_policy::provision(&*provisioned_policy, &base).unwrap();
-
-        let os_policy: Arc<dyn crate::managed_policy::OsPolicySource> =
-            Arc::new(crate::managed_policy::NoOsPolicy);
-        let runtime = GatewayRuntime::new(
-            store.clone(),
-            secrets.clone(),
-            provisioned_policy.clone(),
-            os_policy.clone(),
-        );
-        let chatgpt = Arc::new(
-            crate::chatgpt_runtime::ChatGptRuntime::new(store.clone(), secrets.clone()).unwrap(),
-        );
-        let resolver = Arc::new(crate::resolver::ConfiguredResolver::new(
-            store.clone(),
-            secrets.clone(),
-            runtime.clone(),
-            chatgpt.clone(),
-            provisioned_policy.clone(),
-            os_policy.clone(),
-        ));
-        let state = crate::state::AppState::with_gateway_runtime(
-            tidebreak_core::Config::desktop(directory.path()),
-            store.clone(),
-            resolver,
-            secrets,
-            Arc::new(tidebreak_core::ToolRegistry::new()),
-            tidebreak_core::AgentConfig::default(),
-            uuid::Uuid::new_v4(),
-            runtime,
-            chatgpt,
-            provisioned_policy,
-            os_policy,
-        )
-        .unwrap();
-
-        // Identity first: dispatch resolves through the very runtime the
-        // routes and resolver hold — the cheap pin on the bug class.
-        let dispatch = state.mcp.gateway_endpoints();
-        assert!(
-            std::ptr::eq(
-                Arc::as_ptr(&dispatch) as *const (),
-                Arc::as_ptr(&state.gateway) as *const (),
-            ),
-            "MCP dispatch must share the state's gateway runtime"
-        );
-
-        // And behaviorally: mount an endpoint through the state's MCP
-        // runtime, dispatch one chat's tools/call, mint that chat's
-        // router-facing inference token — one attestation context.
-        let info = state
-            .mcp
-            .replace(crate::mcp_config::McpServersConfig {
-                servers: vec![crate::mcp_config::McpServerDefinition {
-                    name: "tools".to_string(),
-                    command: None,
-                    args: Vec::new(),
-                    env: std::collections::BTreeSet::new(),
-                    env_values: std::collections::BTreeMap::new(),
-                    env_from: Vec::new(),
-                    cwd: None,
-                    url: None,
-                    bearer_token_env: None,
-                    gateway_endpoint: Some("tools".to_string()),
-                    request_timeout_ms: 60_000,
-                    enabled: true,
-                    plugin: None,
-                    launch: None,
-                }],
-            })
-            .await
-            .unwrap();
-        assert_eq!(
-            info.servers[0].health,
-            crate::mcp_config::McpHealth::Healthy
-        );
-
-        let chat = tidebreak_core::id::ChatId::new();
-        let snapshot = state.mcp.snapshot();
-        let tool = snapshot.get("mcp__tools__lookup").unwrap();
-        let ctx = tidebreak_core::ToolCtx::without_private_scratch(chat, None);
-        tool.execute(&ctx, json!({})).await.unwrap();
-        let source = state
-            .gateway
-            .route_token_source()
-            .await
-            .expect("signed-in runtime offers a token source");
-        source.bearer_token_for(Some(chat)).await.unwrap();
-
-        let minted = gateway.minted.lock().unwrap();
-        let chat_mcp = minted
-            .iter()
-            .filter(|(resource, _)| resource == "mcp:tools")
-            .nth(1)
-            .map(|(_, context)| context.clone().unwrap())
-            .expect("the tool call minted a chat-scoped mcp:tools token");
-        let chat_llm = minted
-            .iter()
-            .find(|(resource, _)| resource == "llm")
-            .map(|(_, context)| context.clone().unwrap())
-            .expect("the chat's inference minted an llm token");
-        assert_eq!(
-            chat_mcp, chat_llm,
-            "inference and MCP dispatch must mint in the same attestation context"
-        );
-    }
-
-    /// Mount-by-default, end to end: a reconcile against the entitled apps
-    /// list mounts the endpoint enabled and connected, and a second
-    /// reconcile with unchanged entitlements is a strict no-op — the
-    /// persisted records are untouched, not rewritten to the same shape.
-    #[tokio::test]
-    async fn reconcile_auto_mounts_a_newly_entitled_endpoint_exactly_once() {
-        let address = serve(Arc::new(FakeGateway::default())).await;
-        let base = format!("http://{address}");
-        let (runtime, store, _directory) = signed_in_runtime(&base).await;
-        let mcp = mcp_for(&runtime, &store);
-
-        runtime.reconcile_endpoint_mounts(&mcp).await.unwrap();
-        let info = mcp.info().await;
-        assert_eq!(info.servers.len(), 1);
-        assert_eq!(info.servers[0].definition.name, "tools");
-        assert_eq!(
-            info.servers[0].definition.gateway_endpoint.as_deref(),
-            Some("tools")
-        );
-        assert!(info.servers[0].definition.enabled);
-        assert_eq!(
-            info.servers[0].health,
-            crate::mcp_config::McpHealth::Healthy
-        );
-        assert!(mcp.snapshot().get("mcp__tools__lookup").is_some());
-
-        let records = store.list_connected_apps().await.unwrap();
-        runtime.reconcile_endpoint_mounts(&mcp).await.unwrap();
-        assert_eq!(
-            store.list_connected_apps().await.unwrap(),
-            records,
-            "a reconcile with no new entitlements must not rewrite the records"
-        );
-    }
-
-    /// A failing entitlements fetch degrades to "no reconcile this tick":
-    /// the error surfaces to the caller (which logs and retries) and the
-    /// configuration is untouched.
-    #[tokio::test]
-    async fn a_failing_entitlements_fetch_leaves_the_configuration_untouched() {
-        let gateway = Arc::new(FakeGateway::default());
-        gateway.apps_fail.store(true, Ordering::SeqCst);
-        let address = serve(gateway).await;
-        let base = format!("http://{address}");
-        let (runtime, store, _directory) = signed_in_runtime(&base).await;
-        let mcp = mcp_for(&runtime, &store);
-
-        assert!(runtime.reconcile_endpoint_mounts(&mcp).await.is_err());
-        assert!(mcp.info().await.servers.is_empty());
-        assert!(store.list_connected_apps().await.unwrap().is_empty());
-    }
-
-    /// The sign-in surface on an unmanaged profile exists exactly while a
-    /// pending pairing is parked: it targets the pairing's gateway, commits
-    /// nothing by merely starting, and a dismissal restores the refusal —
-    /// the write-path guard the retired confirmation dialog used to be.
-    #[tokio::test]
-    async fn a_pending_pairing_is_what_sign_in_targets_until_dismissed() {
-        let address = serve(Arc::new(FakeGateway::default())).await;
-        let base = format!("http://{address}/");
-        let directory = tempfile::tempdir().unwrap();
-        let store: Arc<dyn Store> = Arc::new(
-            DbStore::connect(&format!(
-                "sqlite://{}?mode=rwc",
-                directory.path().join("gateway.db").display()
-            ))
-            .await
-            .unwrap(),
-        );
-        let runtime = GatewayRuntime::new(
-            store.clone(),
-            Arc::new(MockSecrets::default()),
-            crate::managed_policy::MemoryProvisionedPolicy::new(),
-            Arc::new(crate::managed_policy::NoOsPolicy),
-        );
-        let mcp = mcp_for(&runtime, &store);
-
-        // Unmanaged with nothing pending: the legible refusal.
-        assert!(runtime.begin_sign_in(mcp.clone()).await.is_err());
-
-        runtime
-            .register_pending_pairing(base.clone(), mcp.clone(), None)
-            .await;
-        let url = runtime.begin_sign_in(mcp.clone()).await.unwrap();
-        assert!(
-            url.starts_with(&format!("{base}oauth/authorize")),
-            "sign-in must target the pending gateway: {url}"
-        );
-        // Starting the flow wrote nothing durable.
-        let policy = runtime.policy().unwrap();
-        assert!(!policy.managed);
-
-        runtime.dismiss_pending_pairing().await;
-        assert_eq!(runtime.pending_pairing_url().await, None);
-        assert_eq!(
-            runtime.status().await.unwrap().sign_in,
-            SignInProgress::Idle
-        );
-        assert!(runtime.begin_sign_in(mcp).await.is_err());
-    }
-
-    #[tokio::test]
-    async fn a_session_for_a_different_gateway_reads_signed_out() {
-        let address = serve(Arc::new(FakeGateway::default())).await;
-        let base = format!("http://{address}");
-        // The profile is managed to one deployment while the stored session
-        // (minted against `base`) belongs to another.
-        let (runtime, _store, _directory) = signed_in_runtime_at(&base, "http://127.0.0.1:9").await;
-
-        let status = runtime.status().await.unwrap();
-        assert!(status.base_url.is_some());
-        assert!(
-            !status.signed_in,
-            "a foreign session must not read signed-in"
-        );
-        assert!(status.account_hint.is_none());
-        assert!(status.installation_id.is_none());
-        assert!(runtime.route_token_source().await.is_none());
-    }
-
-    /// An OS (MDM) authority asserting one gateway, as a device-managed
-    /// profile has.
-    struct OsManaged(String);
-
-    impl crate::managed_policy::OsPolicySource for OsManaged {
-        fn gateway_url(&self) -> Result<Option<String>> {
-            Ok(Some(self.0.clone()))
-        }
-    }
-
-    /// A pure-MDM profile has no stored provider row at all — nothing ever
-    /// wrote one — so reading the status from that row rendered "not
-    /// configured" while sign-in, routing, and mounts all worked. Policy is
-    /// the authority for a managed profile here, exactly as it is for the
-    /// connection itself.
-    #[tokio::test]
-    async fn a_pure_mdm_profile_reads_configured_from_policy_without_a_stored_row() {
-        let address = serve(Arc::new(FakeGateway::default())).await;
-        let base = format!("http://{address}");
-        let directory = tempfile::tempdir().unwrap();
-        let store: Arc<dyn Store> = Arc::new(
-            DbStore::connect(&format!(
-                "sqlite://{}?mode=rwc",
-                directory.path().join("gateway.db").display()
-            ))
-            .await
-            .unwrap(),
-        );
-        let secrets: Arc<dyn SecretProvider> = Arc::new(MockSecrets::default());
-        let credentials: crate::connectors::GatewayCredentials = serde_json::from_value(json!({
-            "base_url": base,
-            "installation_id": "install-1",
-            "user_id": "user-1",
-            "account_hint": "abaas@example.test",
-            "refresh_token": "mg_rt_seed",
-            "access_tokens": {}
-        }))
-        .unwrap();
-        CredentialVault::new(secrets.clone())
-            .save(&credentials)
-            .await
-            .unwrap();
-        let runtime = GatewayRuntime::new(
-            store.clone(),
-            secrets,
-            crate::managed_policy::MemoryProvisionedPolicy::new(),
-            Arc::new(OsManaged(base.clone())),
-        );
-
-        assert!(
-            providers::read_gateway_snapshot(&*store)
-                .await
-                .unwrap()
-                .is_none(),
-            "the fixture must have no stored gateway state for the assertion to mean anything"
-        );
-        let status = runtime.status().await.unwrap();
-        assert!(status.signed_in);
-        assert_eq!(
-            status.base_url.as_deref(),
-            Some(format!("{base}/").as_str())
-        );
-        assert_eq!(status.account_hint.as_deref(), Some("abaas@example.test"));
-    }
-
-    #[tokio::test]
-    async fn a_signed_out_runtime_offers_no_route() {
-        let directory = tempfile::tempdir().unwrap();
-        let store: Arc<dyn Store> = Arc::new(
-            DbStore::connect(&format!(
-                "sqlite://{}?mode=rwc",
-                directory.path().join("gateway.db").display()
-            ))
-            .await
-            .unwrap(),
-        );
-        let secrets: Arc<dyn SecretProvider> = Arc::new(MockSecrets::default());
-        let provisioned = crate::managed_policy::MemoryProvisionedPolicy::new();
-        crate::managed_policy::provision(&*provisioned, "http://127.0.0.1:1").unwrap();
-        let runtime = GatewayRuntime::new(
-            store.clone(),
-            secrets,
-            provisioned,
-            Arc::new(crate::managed_policy::NoOsPolicy),
-        );
-
-        assert!(runtime.route_token_source().await.is_none());
-        let policy = runtime.policy().unwrap();
-        let routes =
-            providers::collect_routes(&*store, &*runtime.secrets, None, None, None, &policy).await;
-        assert!(routes
-            .iter()
-            .all(|route| route.kind != tidebreak_router::RouteKind::ModelGateway));
-    }
-
-    /// The legacy hard cut: an unmanaged profile with a stored additive
-    /// gateway row — and even a leftover signed-in session — has zero
-    /// gateway surface. The row is ignored, never auto-converted to
-    /// managed: lockdown must not be imposed without the pairing consent
-    /// flow, so the remedy is pairing, and the sign-in surface says so.
-    #[tokio::test]
-    async fn an_unmanaged_profile_with_a_legacy_row_has_no_gateway_surface() {
-        struct StaticTokens;
-
-        #[async_trait]
-        impl BearerTokenSource for StaticTokens {
-            async fn bearer_token(&self) -> Result<String> {
-                Ok("mg_at_test".into())
-            }
-        }
-
-        let address = serve(Arc::new(FakeGateway::default())).await;
-        let base = format!("http://{address}");
-        let directory = tempfile::tempdir().unwrap();
-        let store: Arc<dyn Store> = Arc::new(
-            DbStore::connect(&format!(
-                "sqlite://{}?mode=rwc",
-                directory.path().join("gateway.db").display()
-            ))
-            .await
-            .unwrap(),
-        );
-        let secrets: Arc<dyn SecretProvider> = Arc::new(MockSecrets::default());
-        let credentials: crate::connectors::GatewayCredentials = serde_json::from_value(json!({
-            "base_url": base,
-            "installation_id": "install-1",
-            "user_id": "user-1",
-            "refresh_token": "mg_rt_seed",
-            "access_tokens": {}
-        }))
-        .unwrap();
-        CredentialVault::new(secrets.clone())
-            .save(&credentials)
-            .await
-            .unwrap();
-        providers::write_config(
-            &*store,
-            crate::providers::ProviderKind::ModelGateway,
-            &providers::ProviderConfig {
-                enabled: true,
-                base_url: Some(format!("{base}/")),
-                models: vec![CustomModelConfig {
-                    id: "legacy-model".to_string(),
-                    upstream_id: None,
-                    display_name: None,
-                    context_window: 32_768,
-                    max_output_tokens: 4_096,
-                    ..Default::default()
-                }],
-            },
-        )
-        .await
-        .unwrap();
-        let runtime = GatewayRuntime::new(
-            store.clone(),
-            secrets.clone(),
-            crate::managed_policy::MemoryProvisionedPolicy::new(),
-            Arc::new(crate::managed_policy::NoOsPolicy),
-        );
-
-        // The boot cutover warns and ignores the row: no snapshot appears
-        // and the profile stays unmanaged.
-        let policy = runtime.policy().unwrap();
-        providers::retire_legacy_gateway_row(&*store, &policy)
-            .await
-            .unwrap();
-        assert!(providers::read_gateway_snapshot(&*store)
-            .await
-            .unwrap()
-            .is_none());
-        assert!(
-            !policy.managed,
-            "a legacy row must never auto-convert the profile to managed"
-        );
-
-        // Including the machine offer: with no gateway of its own to ask,
-        // an unmanaged profile reads none whatever this one advertises.
-        assert!(runtime.offered_machine().await.url.is_none());
-
-        // The status surface reads no gateway and signed out.
-        let status = runtime.status().await.unwrap();
-        assert!(!status.signed_in);
-        assert!(status.base_url.is_none());
-        assert_eq!(status.model_count, 0);
-
-        // Routing, the picker, and enumeration offer nothing — even with a
-        // token source in hand, the gateway route is not built.
-        assert!(runtime.route_token_source().await.is_none());
-        let tokens: Arc<dyn BearerTokenSource> = Arc::new(StaticTokens);
-        let routes =
-            providers::collect_routes(&*store, &*secrets, Some(tokens), None, None, &policy).await;
-        assert!(routes
-            .iter()
-            .all(|route| route.kind != tidebreak_router::RouteKind::ModelGateway));
-        assert!(providers::catalog_models(&*store, &*secrets, &policy, None)
-            .await
-            .unwrap()
-            .iter()
-            .all(|model| model.policy.provider != crate::providers::ProviderKind::ModelGateway));
-        assert!(providers::list_providers(&*store, &*secrets, &policy, None)
-            .await
-            .unwrap()
-            .iter()
-            .all(|provider| provider.kind != crate::providers::ProviderKind::ModelGateway));
-        assert!(!providers::provider_is_usable(
-            &*store,
-            &*secrets,
-            crate::providers::ProviderKind::ModelGateway,
-            &policy,
-            None
-        )
-        .await
-        .unwrap());
-
-        // The sign-in surface is managed-only, and the refusal names the
-        // remedy.
-        let error = runtime
-            .begin_sign_in(mcp_for(&runtime, &store))
-            .await
-            .err()
-            .unwrap();
-        assert!(
-            error.to_string().contains("pair via your gateway"),
-            "{error}"
-        );
-        assert!(runtime.sync_models().await.is_err());
-        assert!(runtime
-            .apps(&tidebreak_core::OwnerId::local())
-            .await
-            .is_err());
-        assert!(runtime.sign_out().await.is_err());
-    }
-
-    /// The boot cutover's carry-forward: a managed profile — the shape a
-    /// gateway-page pairing produces — keeps the models its row had synced,
-    /// once, and only from the deployment policy actually names.
-    ///
-    /// Nothing re-syncs the entitled set for a reader who is already signed
-    /// in, so dropping this leaves their picker empty until they find the
-    /// refresh button.
-    ///
-    /// The row URL is the VERBATIM form here: the old provider write path
-    /// did not normalize (#935), so a profile that became managed by MDM
-    /// over such a row holds "https://corp.gateway" beside a policy's
-    /// "https://corp.gateway/". Compare deployments as strings instead of
-    /// URLs and that profile silently reaches the picker empty.
-    #[tokio::test]
-    async fn boot_carries_a_managed_rows_snapshot_forward_once() {
-        let directory = tempfile::tempdir().unwrap();
-        let store: Arc<dyn Store> = Arc::new(
-            DbStore::connect(&format!(
-                "sqlite://{}?mode=rwc",
-                directory.path().join("gateway.db").display()
-            ))
-            .await
-            .unwrap(),
-        );
-        let provisioned = crate::managed_policy::MemoryProvisionedPolicy::new();
-        crate::managed_policy::provision(&*provisioned, "https://corp.gateway").unwrap();
-        let legacy_row = |id: &str| providers::ProviderConfig {
-            enabled: true,
-            base_url: Some("https://corp.gateway".to_string()),
-            models: vec![CustomModelConfig {
-                id: id.to_string(),
-                upstream_id: None,
-                display_name: None,
-                context_window: 32_768,
-                max_output_tokens: 4_096,
-                ..Default::default()
-            }],
-        };
-        providers::write_config(
-            &*store,
-            crate::providers::ProviderKind::ModelGateway,
-            &legacy_row("carried-model"),
-        )
-        .await
-        .unwrap();
-
-        let policy =
-            crate::managed_policy::resolve(&*provisioned, &crate::managed_policy::NoOsPolicy)
-                .unwrap();
-        providers::retire_legacy_gateway_row(&*store, &policy)
-            .await
-            .unwrap();
-        let models = providers::gateway_models(&*store, &policy, None)
-            .await
-            .unwrap();
-        assert_eq!(models.len(), 1);
-        assert_eq!(models[0].id, "carried-model");
-
-        // The row is gone once it has been dealt with, so "retired entirely"
-        // is true of the store and not only of the read paths.
-        assert!(
-            providers::read_config(&*store, crate::providers::ProviderKind::ModelGateway)
-                .await
-                .unwrap()
-                .base_url
-                .is_none()
-        );
-
-        // One-shot: once a snapshot exists, a row that reappears never
-        // overwrites it.
-        providers::write_config(
-            &*store,
-            crate::providers::ProviderKind::ModelGateway,
-            &legacy_row("other-model"),
-        )
-        .await
-        .unwrap();
-        providers::retire_legacy_gateway_row(&*store, &policy)
-            .await
-            .unwrap();
-        let models = providers::gateway_models(&*store, &policy, None)
-            .await
-            .unwrap();
-        assert_eq!(models.len(), 1);
-        assert_eq!(models[0].id, "carried-model");
-    }
-
-    /// A row from a deployment the profile is no longer managed by is not
-    /// carried forward: its models describe another gateway's entitlements.
-    /// The row still goes.
-    #[tokio::test]
-    async fn boot_discards_a_snapshot_from_a_foreign_deployment() {
-        let directory = tempfile::tempdir().unwrap();
-        let store: Arc<dyn Store> = Arc::new(
-            DbStore::connect(&format!(
-                "sqlite://{}?mode=rwc",
-                directory.path().join("gateway.db").display()
-            ))
-            .await
-            .unwrap(),
-        );
-        let provisioned = crate::managed_policy::MemoryProvisionedPolicy::new();
-        crate::managed_policy::provision(&*provisioned, "https://corp.gateway").unwrap();
-        providers::write_config(
-            &*store,
-            crate::providers::ProviderKind::ModelGateway,
-            &providers::ProviderConfig {
-                enabled: true,
-                base_url: Some("https://other.gateway".to_string()),
-                models: vec![CustomModelConfig {
-                    id: "foreign-model".to_string(),
-                    upstream_id: None,
-                    display_name: None,
-                    context_window: 32_768,
-                    max_output_tokens: 4_096,
-                    ..Default::default()
-                }],
-            },
-        )
-        .await
-        .unwrap();
-
-        let policy =
-            crate::managed_policy::resolve(&*provisioned, &crate::managed_policy::NoOsPolicy)
-                .unwrap();
-        providers::retire_legacy_gateway_row(&*store, &policy)
-            .await
-            .unwrap();
-        assert!(providers::read_gateway_snapshot(&*store)
-            .await
-            .unwrap()
-            .is_none());
-        assert!(
-            providers::read_config(&*store, crate::providers::ProviderKind::ModelGateway)
-                .await
-                .unwrap()
-                .base_url
-                .is_none()
-        );
-    }
-
-    /// The unmanaged half of the cutover: the row is ignored (never
-    /// converted to managed) and dropped, so the warning naming the remedy
-    /// is a one-time upgrade notice rather than a line on every boot.
-    #[tokio::test]
-    async fn boot_drops_an_unmanaged_legacy_row_without_making_it_managed() {
-        let directory = tempfile::tempdir().unwrap();
-        let store: Arc<dyn Store> = Arc::new(
-            DbStore::connect(&format!(
-                "sqlite://{}?mode=rwc",
-                directory.path().join("gateway.db").display()
-            ))
-            .await
-            .unwrap(),
-        );
-        providers::write_config(
-            &*store,
-            crate::providers::ProviderKind::ModelGateway,
-            &providers::ProviderConfig {
-                enabled: true,
-                base_url: Some("https://corp.gateway".to_string()),
-                models: vec![CustomModelConfig {
-                    id: "legacy-model".to_string(),
-                    upstream_id: None,
-                    display_name: None,
-                    context_window: 32_768,
-                    max_output_tokens: 4_096,
-                    ..Default::default()
-                }],
-            },
-        )
-        .await
-        .unwrap();
-
-        let provisioned = crate::managed_policy::MemoryProvisionedPolicy::new();
-        let policy =
-            crate::managed_policy::resolve(&*provisioned, &crate::managed_policy::NoOsPolicy)
-                .unwrap();
-        providers::retire_legacy_gateway_row(&*store, &policy)
-            .await
-            .unwrap();
-
-        assert!(
-            !policy.managed,
-            "a legacy row must never auto-convert the profile to managed"
-        );
-        assert!(providers::read_gateway_snapshot(&*store)
-            .await
-            .unwrap()
-            .is_none());
-        assert!(
-            providers::read_config(&*store, crate::providers::ProviderKind::ModelGateway)
-                .await
-                .unwrap()
-                .base_url
-                .is_none()
-        );
-    }
-
-    /// A secret store whose reads fail while `fail_reads` is set — the
-    /// transient shape of a keychain read whose ACL no longer matches the
-    /// running binary (denied or pending prompt), as opposed to a confirmed
-    /// absence.
-    #[derive(Default)]
-    struct FlakySecrets {
-        values: MockSecrets,
-        fail_reads: std::sync::atomic::AtomicBool,
-    }
-
-    #[async_trait]
-    impl SecretProvider for FlakySecrets {
-        async fn get_secret(&self, key: &str) -> Result<Option<String>> {
-            if self.fail_reads.load(Ordering::SeqCst) {
-                return Err(AgentError::Secret("keychain read denied".into()));
-            }
-            self.values.get_secret(key).await
-        }
-        async fn set_secret(&self, key: &str, value: &str) -> Result<()> {
-            self.values.set_secret(key, value).await
-        }
-        async fn delete_secret(&self, key: &str) -> Result<()> {
-            self.values.delete_secret(key).await
-        }
-    }
-
-    /// Retire is a one-way door — the session is gone afterwards — so it
-    /// must not open on a transient read error: a keychain read that fails
-    /// (or an unparsable blob, which also errors the load) proves nothing
-    /// about whether the session is superseded. Only a successful read that
-    /// shows a definitive policy mismatch may retire it.
-    #[tokio::test]
-    async fn boot_keeps_the_session_when_the_credential_read_errors() {
-        let secrets = Arc::new(FlakySecrets::default());
-        let credentials: crate::connectors::GatewayCredentials = serde_json::from_value(json!({
-            "base_url": "http://127.0.0.1:1",
-            "installation_id": "install-1",
-            "user_id": "user-1",
-            "refresh_token": "mg_rt_survivor",
-            "access_tokens": {}
-        }))
-        .unwrap();
-        CredentialVault::new(secrets.clone())
-            .save(&credentials)
-            .await
-            .unwrap();
-
-        // Unmanaged policy: a readable session WOULD be retired. With the
-        // read erroring, retire must leave it exactly where it is.
-        let provisioned = crate::managed_policy::MemoryProvisionedPolicy::new();
-        let policy =
-            crate::managed_policy::resolve(&*provisioned, &crate::managed_policy::NoOsPolicy)
-                .unwrap();
-        assert!(!policy.managed);
-        secrets.fail_reads.store(true, Ordering::SeqCst);
-        retire_superseded_gateway_session(secrets.clone(), &policy)
-            .await
-            .unwrap();
-        secrets.fail_reads.store(false, Ordering::SeqCst);
-        assert!(
-            crate::connectors::has_stored_credentials(&*secrets).await,
-            "a transient read error must not retire the session"
-        );
-    }
-
-    /// The session half of the legacy hard cut: an unmanaged profile with a
-    /// session left over from the retired additive mode has no surface that
-    /// could ever revoke it, so boot clears it. Without this the refresh
-    /// token lives in the keychain forever.
-    #[tokio::test]
-    async fn boot_clears_a_gateway_session_left_on_an_unmanaged_profile() {
-        let secrets: Arc<dyn SecretProvider> = Arc::new(MockSecrets::default());
-        let seed = |secrets: Arc<dyn SecretProvider>, base_url: &'static str| async move {
-            let credentials: crate::connectors::GatewayCredentials =
-                serde_json::from_value(json!({
-                    "base_url": base_url,
-                    "installation_id": "install-1",
-                    "user_id": "user-1",
-                    "refresh_token": "mg_rt_zombie",
-                    "access_tokens": {}
-                }))
-                .unwrap();
-            CredentialVault::new(secrets)
-                .save(&credentials)
-                .await
-                .unwrap();
-        };
-        // Nothing listens here: the revoke fails fast and the clear happens
-        // anyway, which is the contract.
-        seed(secrets.clone(), "http://127.0.0.1:1").await;
-
-        let provisioned = crate::managed_policy::MemoryProvisionedPolicy::new();
-        let policy =
-            crate::managed_policy::resolve(&*provisioned, &crate::managed_policy::NoOsPolicy)
-                .unwrap();
-        assert!(!policy.managed);
-        retire_superseded_gateway_session(secrets.clone(), &policy)
-            .await
-            .unwrap();
-        assert!(
-            !crate::connectors::has_stored_credentials(&*secrets).await,
-            "the retired session must not survive boot on an unmanaged profile"
-        );
-
-        // A stored base URL that no longer passes the gateway contract has
-        // no connection to revoke through, so only the unconditional clear
-        // can retire it. Drop that clear and the credential survives here.
-        seed(secrets.clone(), "http://user:pw@stale.example").await;
-        retire_superseded_gateway_session(secrets.clone(), &policy)
-            .await
-            .unwrap();
-        assert!(
-            !crate::connectors::has_stored_credentials(&*secrets).await,
-            "a session whose stored URL cannot be parsed must still be cleared"
-        );
-
-        // A managed profile's session is untouched: it is the credential the
-        // profile actually runs on.
-        seed(secrets.clone(), "https://corp.gateway/").await;
-        crate::managed_policy::provision(&*provisioned, "https://corp.gateway").unwrap();
-        let policy =
-            crate::managed_policy::resolve(&*provisioned, &crate::managed_policy::NoOsPolicy)
-                .unwrap();
-        retire_superseded_gateway_session(secrets.clone(), &policy)
-            .await
-            .unwrap();
-        assert!(crate::connectors::has_stored_credentials(&*secrets).await);
-    }
-
-    /// The managed analogue of the legacy hard cut: an MDM re-point
-    /// supersedes the stored session, and boot is the only surface that will
-    /// ever revoke it at the old deployment.
-    #[tokio::test]
-    async fn boot_retires_the_session_an_mdm_repoint_superseded() {
-        let secrets: Arc<dyn SecretProvider> = Arc::new(MockSecrets::default());
-        let old_gateway = Arc::new(FakeGateway::default());
-        let old_base = format!("http://{}", serve(old_gateway.clone()).await);
-        let credentials: crate::connectors::GatewayCredentials = serde_json::from_value(json!({
-            "base_url": old_base,
-            "installation_id": "install-1",
-            "user_id": "user-1",
-            "refresh_token": "mg_rt_zombie",
-            "access_tokens": {}
-        }))
-        .unwrap();
-        CredentialVault::new(secrets.clone())
-            .save(&credentials)
-            .await
-            .unwrap();
-        let provisioned = crate::managed_policy::MemoryProvisionedPolicy::new();
-        crate::managed_policy::provision(&*provisioned, "https://corp-new.gateway").unwrap();
-        let mut policy =
-            crate::managed_policy::resolve(&*provisioned, &crate::managed_policy::NoOsPolicy)
-                .unwrap();
-        assert!(policy.managed);
-
-        // A managed policy with no usable URL is misconfiguration, not a
-        // re-point: the session is left for the repaired policy to judge.
-        policy.gateway_url = None;
-        retire_superseded_gateway_session(secrets.clone(), &policy)
-            .await
-            .unwrap();
-        assert!(crate::connectors::has_stored_credentials(&*secrets).await);
-
-        policy.gateway_url = Some("https://corp-new.gateway".into());
-        retire_superseded_gateway_session(secrets.clone(), &policy)
-            .await
-            .unwrap();
-        assert!(
-            !crate::connectors::has_stored_credentials(&*secrets).await,
-            "the superseded session must not survive the re-point"
-        );
-        // Revoked at the old deployment, with the superseded refresh token.
-        assert_eq!(
-            old_gateway.revoked.lock().unwrap().as_slice(),
-            ["mg_rt_zombie"]
-        );
-    }
-
-    /// The durability the policy file exists for: a profile below the
-    /// migration pin is deleted and rebuilt on first boot, and the
-    /// provisioned policy — a sidecar file in the data directory — must
-    /// survive that. The session it authorizes is then retained: before the
-    /// policy lived in the file, the reset resolved the profile unmanaged and
-    /// boot retired the session, forcing a full re-pair.
-    #[tokio::test]
-    async fn a_schema_epoch_reset_keeps_the_policy_and_its_session() {
-        let directory = tempfile::tempdir().unwrap();
-        let data_dir = directory.path();
-        let provisioned = crate::managed_policy::ProvisionedPolicyFile::in_data_dir(data_dir);
-        crate::managed_policy::provision(&provisioned, "https://corp.gateway").unwrap();
-
-        // The profile's database, holding a chat so its loss is observable,
-        // and the gateway session the policy authorizes.
-        let store = DbStore::connect(&format!(
-            "sqlite://{}?mode=rwc",
-            data_dir.join("tidebreak.db").display()
-        ))
-        .await
-        .unwrap();
-        store
-            .create_chat(&tidebreak_core::Chat {
-                id: tidebreak_core::ChatId::new(),
-                project_id: None,
-                title: Some("lost to the reset".to_owned()),
-                model: None,
-                reasoning_effort: None,
-                permission_mode: None,
-                network_policy: Default::default(),
-                attachment_revision: 0,
-                root_attachments: Vec::new(),
-                created_at: chrono::Utc::now(),
-            })
-            .await
-            .unwrap();
-        // An explicit close, not a drop: the reset below deletes the SQLite
-        // files, which Windows refuses while a handle is open.
-        store.close().await.unwrap();
-        let secrets: Arc<dyn SecretProvider> = Arc::new(MockSecrets::default());
-        let credentials: crate::connectors::GatewayCredentials = serde_json::from_value(json!({
-            "base_url": "https://corp.gateway/",
-            "installation_id": "install-1",
-            "user_id": "user-1",
-            "refresh_token": "mg_rt_seed",
-            "access_tokens": {}
-        }))
-        .unwrap();
-        CredentialVault::new(secrets.clone())
-            .save(&credentials)
-            .await
-            .unwrap();
-
-        // The epoch reset, as desktop_schema performs it: the database files
-        // are deleted; everything else in the data directory stays.
-        std::fs::remove_file(data_dir.join("tidebreak.db")).unwrap();
-        assert!(!data_dir.join("tidebreak.db").exists());
-
-        // Resolution needs no store at all now: the policy survives, still
-        // managed to the same gateway, and the retire step — boot's session
-        // cleanup — keeps the session it stands behind.
-        let policy =
-            crate::managed_policy::resolve(&provisioned, &crate::managed_policy::NoOsPolicy)
-                .unwrap();
-        assert!(policy.managed && !policy.misconfigured);
-        assert_eq!(policy.gateway_url.as_deref(), Some("https://corp.gateway/"));
-        retire_superseded_gateway_session(secrets.clone(), &policy)
-            .await
-            .unwrap();
-        assert!(
-            crate::connectors::has_stored_credentials(&*secrets).await,
-            "a policy that survives the reset must keep the session it authorizes"
-        );
-    }
-
-    /// A completed sign-in supersedes whatever session is stored — possibly
-    /// one minted at a different deployment, after a re-point. Its refresh
-    /// token is revoked at its own gateway before the overwrite orphans it.
-    #[tokio::test]
-    async fn signing_in_revokes_the_superseded_session_at_its_own_gateway() {
-        let old_gateway = Arc::new(FakeGateway::default());
-        let old_base = format!("http://{}", serve(old_gateway.clone()).await);
-        let new_gateway = Arc::new(FakeGateway::default());
-        let new_base = format!("http://{}", serve(new_gateway.clone()).await);
-        let secrets: Arc<dyn SecretProvider> = Arc::new(MockSecrets::default());
-        let superseded: crate::connectors::GatewayCredentials = serde_json::from_value(json!({
-            "base_url": old_base,
-            "installation_id": "install-1",
-            "user_id": "user-1",
-            "refresh_token": "mg_rt_zombie",
-            "access_tokens": {}
-        }))
-        .unwrap();
-        CredentialVault::new(secrets.clone())
-            .save(&superseded)
-            .await
-            .unwrap();
-
-        let connection = GatewayConnection::new(
-            GatewayAuth::new(GatewayAuthConfig::new(&new_base).unwrap()).unwrap(),
-            CredentialVault::new(secrets.clone()),
-        );
-        connection
-            .store_session(&crate::connectors::AuthorizedSession {
-                meta: crate::connectors::GatewayMeta {
-                    api_version: "1".into(),
-                    installation_id: "install-2".into(),
-                    gateway_version: "1".into(),
-                    public_url: new_base.clone(),
-                    auth_mode: "oauth".into(),
-                    tidebreak_machine_url: None,
-                    surfaces: None,
-                },
-                identity: crate::connectors::GatewayIdentity {
-                    user_id: "user-2".into(),
-                    email: Some("user@example.test".into()),
-                    display_name: None,
-                    session_id: "session-2".into(),
-                    installation_id: "install-2".into(),
-                },
-                tokens: crate::connectors::TokenSet {
-                    access_token: "mg_at_fresh".into(),
-                    refresh_token: "mg_rt_fresh".into(),
-                    expires_at_unix: u64::MAX,
-                    scope: "models:read inference:invoke".into(),
-                    resource: "control".into(),
-                    installation_id: "install-2".into(),
-                },
-            })
-            .await
-            .unwrap();
-
-        // The revoke went to the superseded session's own deployment — not
-        // the connection's — and the new session is stored for the new one.
-        assert_eq!(
-            old_gateway.revoked.lock().unwrap().as_slice(),
-            ["mg_rt_zombie"]
-        );
-        assert!(new_gateway.revoked.lock().unwrap().is_empty());
-        assert!(crate::connectors::has_stored_credentials_for(&*secrets, &new_base).await);
-    }
-}
